### 2019-01-25 da5bc9a
Added missing 'References' section to JOSS paper.
