+++ joss_paper/paper.md
@@ -56,3 +56,6 @@ EV would also like to thank Chris Blake, Colin Jacobs and Theo Steininger for fr
 Parts of this research were supported by the Australian Research Council Centre of Excellence for All Sky Astrophysics in 3 Dimensions (ASTRO 3D), through project number CE170100013.
 Parts of this work were performed on the OzSTAR national facility at Swinburne University of Technology. OzSTAR is funded by Swinburne University of Technology and the National Collaborative Research Infrastructure Strategy (NCRIS).
 
+
+# References
+
