### 2019-01-28 490aed7
Added PRISM paper citation to README and docs.
