### 2019-01-28 11fb59e
Added PRISM paper citation to README and docs.
