### 2019-01-28 6a81512
Added PRISM paper citation to README and docs.
