+++ README.rst
@@ -93,6 +93,38 @@ See `online docs`_ for a documented explanation on this example.
     pipe.project()
 
 
+Community guidelines
+====================
+*PRISM* is an open-source and free-to-use software package (and it always will be), provided under the `BSD-3 license`_.
+Users are highly encouraged to make contributions to the package or request new features by opening a `GitHub issue`_.
+If you would like to contribute to the package, but do not know what, then there are quite a few ToDos in the code that may give you some inspiration.
+As with contributions, if you find a problem or issue with *PRISM*, please do not hesitate to open a `GitHub issue`_ about it.
+Keep in mind that issues solely related to the `v1.0.x` branch are given a lower priority.
+
+And, finally, if you use *PRISM* as part of your workflow in a scientific publication, please consider citing the *PRISM* pipeline paper (and starring the `repository`_):
+
+::
+
+    @ARTICLE{2019arXiv190108725V,
+       author = {{van der Velden}, E. and {Duffy}, A.~R. and {Croton}, D. and 
+    	{Mutch}, S.~J. and {Sinha}, M.},
+        title = "{Model dispersion with PRISM; an alternative to MCMC for rapid analysis of models}",
+      journal = {arXiv e-prints},
+    archivePrefix = "arXiv",
+       eprint = {1901.08725},
+     primaryClass = "astro-ph.IM",
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics},
+         year = 2019,
+        month = jan,
+       adsurl = {http://adsabs.harvard.edu/abs/2019arXiv190108725V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+
+
+.. _BSD-3 license: https://github.com/1313e/PRISM/raw/master/LICENSE
+.. _GitHub issue: https://github.com/1313e/PRISM/issues
+
+
 .. |PyPI| image:: https://img.shields.io/badge/PyPI-v1.0.0-blue.svg
     :target: https://pypi.python.org/pypi/prism/v1.0.0
     :alt: PyPI - Latest v1.0.x Release
