+++ README.rst
@@ -134,10 +134,10 @@ And, finally, if you use *PRISM* as part of your workflow in a scientific public
 .. |License| image:: https://img.shields.io/pypi/l/prism.svg?colorB=blue&label=License
     :target: https://github.com/1313e/PRISM/raw/v1.0.x/LICENSE
     :alt: PyPI - License
-.. |Travis| image:: https://img.shields.io/travis/com/1313e/PRISM/v1.0.x.svg?logo=travis&label=Travis%20CI
+.. |Travis| image:: https://img.shields.io/travis/com/1313e/PRISM/v1.0.x.svg?logo=travis&logoColor=white&label=Travis%20CI
     :target: https://travis-ci.com/1313e/PRISM
     :alt: Travis CI - Build Status
-.. |AppVeyor| image:: https://img.shields.io/appveyor/ci/1313e/PRISM/v1.0.x.svg?logo=appveyor&label=AppVeyor
+.. |AppVeyor| image:: https://img.shields.io/appveyor/ci/1313e/PRISM/v1.0.x.svg?logo=appveyor&logoColor=white&label=AppVeyor
     :target: https://ci.appveyor.com/project/1313e/PRISM/branch/v1.0.x
     :alt: AppVeyor - Build Status
 .. |ReadTheDocs| image:: https://img.shields.io/readthedocs/prism-tool/v1.0.x.svg?logo=read%20the%20docs&logoColor=white&label=Docs
