### 2019-01-29 630004c
Fixed a typo in the README and docs.
