### 2019-01-29 d3650a3
Fixed a typo in the README and docs.
