### 2019-01-29 32f54ed
Fixed a typo in the README and docs.
