### 2019-01-31 393c514
Fixed styling of PRISM paper citation in the docs.
