### 2019-01-31 6e066e4
Fixed styling of PRISM paper citation in the docs.
