+++ docs/source/community_guidelines.rst
@@ -10,7 +10,8 @@ As with contributions, if you find a problem or issue with *PRISM*, please do no
 
 And, finally, if you use *PRISM* as part of your workflow in a scientific publication, please consider citing the *PRISM* pipeline paper (and starring the `repository`_):
 
-.. note::
+.. code-block:: none
+    
     @ARTICLE{2019arXiv190108725V,
        author = {{van der Velden}, E. and {Duffy}, A.~R. and {Croton}, D. and 
     	{Mutch}, S.~J. and {Sinha}, M.},
