### 2019-01-31 9024642
Removed additions list as the v1.0.x branch will not be actively maintained.
