+++ docs/source/community_guidelines.rst
@@ -36,8 +36,3 @@ License
 -------
 .. literalinclude:: ../../LICENSE
     :language: none
-
-
-Additions
----------
-.. include:: additions.inc
