### 2019-02-03 #1200 @arokem
@fonnesbeck : would you be willing to review this article for [JOSS](http://joss.theoj.org/)?
