### 2019-02-04 #1229 @whedon
[REVIEW]: Model dispersion with PRISM; an alternative to MCMC for rapid analysis of models
**Submitting author:** @1313e (<a href="http://orcid.org/0000-0002-1559-9832">Ellert van der Velden</a>)
**Repository:** <a href="https://github.com/1313e/PRISM" target ="_blank">https://github.com/1313e/PRISM</a>
**Version:** v1.0.0
