### 2019-02-04 #1229 @1313e
As mentioned in the blog here (http://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing),
it must be clear to all parties involved that a JOSS submission is associated with an AAS Journals submission.
Given that there is currently (as far as I know) not a defined way of doing so, I would like to mention here that there is an AAS submission associated with this JOSS submission.

@fonnesbeck @arokem Do you need the AAS Journals submission reference for this by any chance?
