### 2019-02-06 4e955ba
Added the _write_data_idx()-method to the Emulator class for easier access.
