### 2019-02-06 c09c775
Made some changes to the community guidelines. Added special RST role for referring to PEPs.
