+++ docs/source/community_guidelines.rst
@@ -8,7 +8,7 @@ Users are highly encouraged to make contributions to the package or request new 
 If you would like to contribute to the package, but do not know what, then there are quite a few ToDos in the code that may give you some inspiration.
 As with contributions, if you find a problem or issue with *PRISM*, please do not hesitate to open a `GitHub issue`_ about it.
 
-And, finally, if you use *PRISM* as part of your workflow in a scientific publication, please consider citing the *PRISM* pipeline paper (and starring the `repository`_):
+And, finally, if you use *PRISM* as part of your workflow in a scientific publication, please consider including an acknowledgement like *"Parts of the results in this work were derived using the PRISM Python package."* and citing the *PRISM* pipeline paper:
 
 .. code-block:: none
     
@@ -27,7 +27,7 @@ And, finally, if you use *PRISM* as part of your workflow in a scientific public
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
     }
 
-.. _repository: https://github.com/1313e/PRISM
+
 .. _GitHub issue: https://github.com/1313e/PRISM/issues
 
 
