### 2019-02-06 4c095fc
Made some changes to the community guidelines. Added special RST role for referring to PEPs.
