### 2019-02-06 a446d49
Made some changes to the community guidelines. Added special RST role for referring to PEPs.
