### 2019-02-07 #1229 @1313e
@arfon May I btw mention that AAS Journals has no system in place at all that a submission is being accompanied by a JOSS submission?
I simply mentioned it in the extra notes to the editor, but I have not heard anything back yet.
