### 2019-02-07 #1229 @arfon
> I simply mentioned it in the extra notes to the editor, but I have not heard anything back yet.

Yes, I think this is still a work in progress.

/ cc @crawfordsm who is the software editor for AAS so he has some visibility into this review too.
