### 2019-02-07 #1229 @crawfordsm
Apologies, unfortunately, I did not see a note about this on the AAS side.   I'll have to follow up about that, but thanks for letting me know!    As this is the first paper, it will be great to use as an example of the process.
