### 2019-02-07 #1229 @1313e
> Apologies, unfortunately, I did not see a note about this on the AAS side. I'll have to follow up about that, but thanks for letting me know! As this is the first paper, it will be great to use as an example of the process.

Do you, by any chance, need the submission reference, to make it easier to find which paper it actually is that I submitted to AAS Journals?
