### 2019-02-07 #1229 @crawfordsm
It's okay, I have the information.   I am the Scientific Editor for the paper, so I can monitor it here as well as on the AAS side.   I realize now that I had just overlooked the information.   As of right now, we do not have a formal process for this, but appreciate your patience as we figure out the details.
