### 2019-02-10 #1229 @arokem
Hi @arfon: thanks for the heads up about this. Do I understand correctly that the process here is essentially the same as for other JOSS papers? That is, review and acceptance here are not tied in any way to the AAS paper. Is that correct?
