### 2019-02-10 #1229 @arfon
> Hi @arfon: thanks for the heads up about this. Do I understand correctly that the process here is essentially the same as for other JOSS papers? That is, review and acceptance here are not tied in any way to the AAS paper. Is that correct?

That's _mostly_ correct. We may want to hold this paper back from final acceptance until we know what the DOI for the AAS paper will be (so we can cite it).
