### 2019-02-10 #1229 @arokem
Understood! OK - @fonnesbeck - if this all sounds agreeable to you, please do go ahead with your review. Once the review process here is over, we can circle back and see what we need to do. Thanks!
