### 2019-02-11 #1229 @1313e
> That's _mostly_ correct. We may want to hold this paper back from final acceptance until we know what the DOI for the AAS paper will be (so we can cite it).

@arfon That would be required anyways, given that I am citing the AAS Journals submission for the results.
I felt that putting them into the JOSS paper would be a bit pointless, given that they are already in the AAS submission and it would never fit into 1k words.

I hope that is okay btw, as I couldn't find any guidelines on this.
