### 2019-02-11 #1229 @arfon
> I felt that putting them into the JOSS paper would be a bit pointless, given that they are already in the AAS submission and it would never fit into 1k words.

:+1: yes that's right. We also don't allow scientific results to be published in JOSS :-)

> I hope that is okay btw, as I couldn't find any guidelines on this.

I think you've got this right. @crawfordsm and I are in the process of drafting some guidelines that should assist future authors.
