### 2019-02-14 0a1c776
Made some small improvements to the overview that the get_info()-function prints.
