### 2019-02-14 a86de60
Changed the interpolation of the projections to use normalized grids instead of scaled grids. This fixes a bug where if the difference in parameter range width was too large (like [0, 100] and [0, 1]), the interpolation would not be executed properly. This also improves the quality of the interpolation in general, as the interpolated projection grid has the same dimension lengths. Fixed a problem with the 2D projections where the axis ranges would not be set properly if an estimate line was too close to the parameter bounds.
