### 2019-02-14 0c73ae6
Set version to 1.0.1.dev0, as I apparently had forgotten to do that several commits ago.
