### 2019-02-14 d173f38
Removed a few unnecessary for-loops. Replaced a few cases of using copy() with deepcopy(). Replaced a single case of individual MPI send/recv by a gather(). Replaced many cases with Python 3 iterable unpacking.
