### 2019-02-15 177672e
Small change to the position of the auxiliary characters list and mentioned it in the docstring of convert_str_seq().
