### 2019-02-15 25ed753
Small change to the position of the auxiliary characters list and mentioned it in the docstring of convert_str_seq().
