### 2019-02-15 #1229 @fonnesbeck
@1313e the test suite fails on both Linux and macOS for me. `pytest` does not like the arguments:

    pytest: error: unrecognized arguments: --mpl --pep8
