### 2019-02-15 1313e/PRISM#7 @fonnesbeck
README example hangs on projection creation.
I've installed PRISM on both a Debian Linux system and a macOS machine successfully, however executing the example shown in the README file does not run to completion. Specifically, it hangs at the projection creation step with 0% completed.

Here is a screen shot of what I am seeing:

![image](https://user-images.githubusercontent.com/81476/52885686-a99b4b80-3137-11e9-819b-028ea4339b3e.png)

I assume the example is meant to run quickly, given how small it is.

In both cases PRISM was installed in a conda virtualenv with Python 3.7.2, using pip on the cloned master branch.
