### 2019-02-15 #1229 @fonnesbeck
I'm having problems running the example model. See 1313e/PRISM#7
