### 2019-02-16 #1229 @1313e
> @1313e the test suite fails on both Linux and macOS for me. `pytest` does not like the arguments:
>
> ```
> pytest: error: unrecognized arguments: --mpl --pep8
> ```

@fonnesbeck You will have to install all requirements in `requirements_dev.txt` before running the pytests.
