### 2019-02-16 47ff986
Changed default value of criterion to None.
