### 2019-02-16 a66a040
Changed default value of criterion to None.
