+++ prism/data/prism.txt
@@ -5,7 +5,7 @@ base_eval_sam       : 800                   # Base number for growth in number o
 sigma               : 0.8                   # Gaussian sigma/standard deviation (only required if method == 'gaussian')
 l_corr              : 0.3                   # Gaussian correlation length(s)
 impl_cut            : [0.0, 4.0, 3.8, 3.5]  # List of implausibility cut-off values
-criterion           : 'multi'               # Criterion for constructing LHDs
+criterion           : None                  # Criterion for constructing LHDs
 method              : 'full'                # Method used for constructing the emulator
 use_regr_cov        : False                 # Use regression covariance
 poly_order          : 3                     # Polynomial order for regression
