### 2019-02-16 1913c68
Changed default value of criterion to None.
