### 2019-02-16 1313e/PRISM#7 @1313e
@fonnesbeck Thanks for your report.
I just tried this on a Windows machine using Python 3.7.2 and simply installing PRISM from the master branch.
The projections are created for me with no problem.

However, for the 2D projections, it tries to create a large Latin-Hypercube that satisfies a bunch of criteria (as is done by default in PRISM) and this causes it take a very long time to create one.
Guess it is finally time to relax that condition a bit.

I just changed the default value of the associated parameter on all branches.
Can you try to run the example again?
