### 2019-02-16 #1229 @fonnesbeck
Yes, at the moment the install instructions do not mention `requirements_dev`. Thanks.
