### 2019-02-17 20e1c57
Added note about requirements_dev.txt being required for running pytests.
