### 2019-02-17 75d5a77
Added note about requirements_dev.txt being required for running pytests.
