### 2019-02-17 #1229 @1313e
> Yes, at the moment the install instructions do not mention `requirements_dev`. Thanks.

@fonnesbeck Alright, I have added it to the README.
I have also "fixed" 1313e/PRISM#7.
