### 2019-02-17 1313e/PRISM#7 @fonnesbeck
OK, it appears to run now (about 120s/hypercube) on my MacBook Pro. Thanks.
