### 2019-02-17 1313e/PRISM#8 @fonnesbeck
Test failure on Linux
I've installed all the dependencies in `requirements_dev` but running `pytest ./prism/tests` results in a failure:

```
(dev) fonnesbeck@penguin:~/repos/PRISM (master)$ pytest prism/tests/
/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pep8.py:110: FutureWarning: Possible nested set at position 1
  EXTRANEOUS_WHITESPACE_REGEX = re.compile(r'[[({] | []}),;:]')
================================ test session starts =================================
platform linux -- Python 3.7.1, pytest-4.2.0, py-1.7.0, pluggy-0.8.1 -- /home/fonnesbeck/anaconda3/envs/dev/bin/python
cachedir: .pytest_cache
Matplotlib: 3.0.2
Freetype: 2.9.1
rootdir: /home/fonnesbeck/repos/PRISM, inifile: setup.cfg
plugins: pep8-1.0.6, mpl-0.10, cov-2.6.1
INTERNALERROR> Traceback (most recent call last):
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/_pytest/main.py", line 203, in wrap_session
INTERNALERROR>     session.exitstatus = doit(config, session) or 0
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/_pytest/main.py", line 242, in _main
INTERNALERROR>     config.hook.pytest_collection(session=session)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/hooks.py", line 284, in __call__
INTERNALERROR>     return self._hookexec(self, self.get_hookimpls(), kwargs)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/manager.py", line 68, in _hookexec
INTERNALERROR>     return self._inner_hookexec(hook, methods, kwargs)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/manager.py", line 62, in <lambda>
INTERNALERROR>     firstresult=hook.spec.opts.get("firstresult") if hook.spec else False,
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/callers.py", line 208, in _multicall
INTERNALERROR>     return outcome.get_result()
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/callers.py", line 80, in get_result
INTERNALERROR>     raise ex[1].with_traceback(ex[2])
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/pluggy/callers.py", line 182, in _multicall
INTERNALERROR>     next(gen)  # first yield
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/_pytest/warnings.py", line 149, in pytest_collection
INTERNALERROR>     config=config, ihook=config.hook, when="collect", item=None
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/contextlib.py", line 112, in __enter__
INTERNALERROR>     return next(self.gen)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/_pytest/warnings.py", line 87, in catch_warnings_for_item
INTERNALERROR>     _setoption(warnings, arg)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/site-packages/_pytest/warnings.py", line 26, in _setoption
INTERNALERROR>     category = wmod._getcategory(category)
INTERNALERROR>   File "/home/fonnesbeck/anaconda3/envs/dev/lib/python3.7/warnings.py", line 255, in _getcategory
INTERNALERROR>     raise _OptionError("invalid module name: %r" % (module,)) from None
INTERNALERROR> warnings._OptionError: invalid module name: 'prism._internal'
```

Running Py 3.7 (Anaconda) on Debian Linux
