### 2019-02-17 #1229 @fonnesbeck
Tests now run after installing the dev requirements, but [the tests fail](https://github.com/1313e/PRISM/issues/8).
