### 2019-02-17 #1229 @fonnesbeck
The submission satisfies all of the evaluation criteria. Once the issue above is addressed, I'm happy to recommend ACCEPT.
