### 2019-02-18 85b9d0d
Made sure that PRISM is now installed before attempting to run pytests. Added a small paragraph to the docs and README about running pytests on PRISM.
