+++ .appveyor.yml
@@ -50,7 +50,6 @@ install:
     - "check-manifest"
     - "python setup.py sdist bdist_wheel"
     - "twine check dist/*"
-    - "pip install -e ."
     - "python -c \"import prism; prism.get_info()\""
 
 build: false
