+++ setup.cfg
@@ -4,8 +4,6 @@ universal=0
 [tool:pytest]
 addopts = --mpl --pep8 -v
 testpaths = ./prism
-filterwarnings = 
-  ignore::prism._internal.RequestWarning
 
 [coverage:run]
 parallel = true
