### 2019-02-18 633dcbc
Added missing 'HOME' folder for installation directory path.
