### 2019-02-18 d868c1a
Added missing 'HOME' folder for installation directory path.
