### 2019-02-18 6bb3ac5
Added missing 'HOME' folder for installation directory path.
