### 2019-02-18 4578cc2
Changed many paths to specifically point to the v1.0.x branch. Ensured that the v1.0.x branch now also only deploys if the commit was tagged.
