### 2019-02-18 1313e/PRISM#8 @1313e
@fonnesbeck Ah, I see what is happening here.
There is one very specific function for which PRISM needs to be installed before it can be executed (`prism.get_info()`), or at least its metadata must be available.
Therefore, I always ran the tests with PRISM already installed, as installing is in my opinion part of testing anyways.
I have made changes to the `requirements_dev.txt` file to make sure that it installs PRISM as well.
That also ensures that all normal requirements of PRISM are installed (which I had forgotten apparently).

Also, as a note, if you want to run all tests in PRISM, simply execute `pytest` (with no additional arguments) in the root of the repository.
