### 2019-02-18 #1229 @1313e
@fonnesbeck I have solved the problem with 1313e/PRISM#8, and also improved the README and docs on the topic of running the pytests.

@arokem Do you have an idea how I can fix the formatting error that whedon gives when compiling the `paper.md`?
It formats my last name incorrectly in the footer of every page (I already noticed that immediately after submission, but had not asked about it yet).
