### 2019-02-18 b2a1db8
Added forgotten specification of using the v1.0 of PRISM when installing from PyPI.
