+++ docs/source/user/getting_started.rst
@@ -12,7 +12,7 @@ Installation
 
 or by installing it directly from `PyPI`_ with::
 
-    $ pip install prism
+    $ pip install prism==1.0
 
 *PRISM* can now be imported as a package with :pycode:`import prism`.
 For using *PRISM* in MPI, ``mpi4py >= 3.0.0`` is required (not installed automatically).
