### 2019-02-18 d08f4cb
No longer force pip to be v18.1.
