### 2019-02-18 97f6b21
No longer force pip to be v18.1.
