+++ .appveyor.yml
@@ -31,7 +31,7 @@ install:
     - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
 
     # Upgrade pip
-    - "python -m pip install --user --upgrade pip==18.1"
+    - "python -m pip install --user --upgrade pip"
 
     # Install PRISM
     - "pip install -r requirements_dev.txt"
