### 2019-02-18 ca42a96
Make sure that pip, setuptools and wheel are at their latest versions when testing.
