### 2019-02-18 7c3b937
Make sure that pip, setuptools and wheel are at their latest versions when testing.
