### 2019-02-18 #1229 @fonnesbeck
I am now able to run the test suite without issue. There are no other outstanding problems, so I recommend ACCEPT. cc @arfon
