### 2019-02-18 #1229 @arokem
@1313e

Two comments on the manuscript:

1. The first is that the link to the software is not necessary. As you can see in the rendered PDF, there will be a link to your GitHub repo on the first page of the manuscript. This is not a required fix.

2. Could you please add the DOI for the Raftery et al. paper? (DOI: 10.1080/01621459.1995.10476529). This one is required.
