### 2019-02-18 72458f7
Added DOI to Raftery 1995.
