+++ joss_paper/paper.bib
@@ -1,4 +1,4 @@
-@book{BLA,
+﻿@book{BLA,
 	author				=	{M. Goldstein and D. Wooff},
 	title 				=	"{Bayes Linear Statistics: Theory and Methods}",
 	edition				=	{1},
@@ -80,7 +80,8 @@ doi="10.1186/s12918-017-0484-3"
  publisher = {[American Statistical Association, Taylor & Francis, Ltd.]},
  title = "{Inference from a Deterministic Population Dynamics Model for Bowhead Whales}",
  volume = {90},
- year = {1995}
+ year = {1995},
+ doi = {10.1080/01621459.1995.10476529}
 }
 
 @InProceedings{Craig96,
