### 2019-02-18 #1229 @1313e
> 2. Could you please add the DOI for the Raftery et al. paper? (DOI: 10.1080/01621459.1995.10476529). This one is required.

Alright, fixed that.

> @arokem - let me know when you're ready to accept this - I'll have to manually process this one to deal with Whedon's poor handling of `van der Velden` as a last name.

I guess we will have to wait with accepting the paper until the AAS Journals submission has been reviewed, right?
