### 2019-02-18 #1229 @arfon
> I guess we will have to wait with accepting the paper until the AAS Journals submission has been reviewed, right?

Yes, I think that's right. @crawfordsm - does that sound about right to you?
