### 2019-02-19 #1229 @arokem
OK - I believe that this manuscript is ready to be accepted from our end.

@arfon: should @1313e create a Zenodo archive with the code as it is now, or wait until the AAS journal review is also resolved?
