### 2019-02-19 074048b
Made many changes to the logging system such that now only logging messages originating from PRISM are saved to the logging file, while all other logging messages are handled by the root logger (or any other custom logger somebody might have initialized). All regular logging solely done by PRISM can now be controlled using the do_logging property.
