### 2019-02-19 #1229 @arfon
> @arfon: should @1313e create a Zenodo archive with the code as it is now, or wait until the AAS journal review is also resolved?

I think it's OK to create the archive now.
