### 2019-02-20 #1229 @1313e
So, I guess all that is remaining is linking the JOSS submission with the AAS Journals submission (which is necessary anyway for the last paragraph) and manually adjust the PDF to have my last name spelled correctly?

Btw, I am actually wondering why whedon is writing my last name this weird.
I have seen my name being spelled incorrectly in many different ways before, as most other countries do not understand that a last name can contain multiple words which should not be capitalized, but I have never seen it being spelled as "Van Velden" before.

Fun fact: I have seen no system thus far (besides those in my home country obviously) that can spell it correctly and still understand that my name should be alphabetically sorted on "Velden" rather than "van".
For my name it does not matter that much given how close the two words are together alphabetically speaking, but there are many cases in which this is not true.
