### 2019-02-20 0ee4f6b
Introduced an experimental 'hickle'-based backup system for the call_model()-method of the ModelLink class. The new _make_backup()-method can be called in the call_model()-method to create a backup of all arguments passed to it, plus the variables that were passed to the call_model()-method itself. The new _read_backup()-method can be used to read the backup files back in. Written pytests for the new experimental features. Added hickle to the requirements. Introduced the FeatureWarning, which is raised whenever an experimental feature is used. The RequestWarning now inherits from UserWarning r
[…]
