+++ requirements.txt
@@ -1,5 +1,6 @@
 e13tools>=0.4.7a1
 h5py>=2.8.0
+hickle>=3.3.3
 matplotlib>=2.2.0
 mlxtend>=0.9.1
 numpy>=1.12.0

[…]
