### 2019-02-21 4aa849a
Fixed a small typo in _read_backup(). Made the one-liner explanation in _make_backup() part of the documentation.
