### 2019-02-22 647e565
The test_subclass()-function now additionally checks if the provided subclass has overridden call_model() and get_md_var()-methods that take the correct input arguments.
