### 2019-02-22 04aa48d
The test_subclass()-function now also checks if the call_model() and get_md_var()-method take no other non-optional input arguments.
