### 2019-02-22 b2bd7c6
Added a sorter for the data_idx parts to ensure that they are put together properly (string sorting works differently from integer sorting).
