### 2019-02-22 16321df
Added a sorter for the data_idx parts to ensure that they are put together properly (string sorting works differently from integer sorting).
