### 2019-02-22 #1229 @arokem
I must be missing something. Isn't your last name already spelled correctly [here](https://github.com/openjournals/joss-papers/blob/joss.01229/joss.01229/10.21105.joss.01229.pdf)?
