### 2019-02-23 #1229 @arfon
Just to (re)confirm. When we finally accept this paper, I will fix this manually. The library we use for handling firstnames/lastnames can't handle `van der` so I have to fix papers by hand when this comes up.
