### 2019-02-25 cdefa52
Moved the data_idx keys sorting to the branch where there are more than 1 (as a single data_idx cannot be sorted).
