### 2019-02-25 4b5384c
Moved the data_idx keys sorting to the branch where there are more than 1 (as a single data_idx cannot be sorted).
