### 2019-02-25 5674ed0
Changed the required version for hickle to 3.4.0. Changed the hickle dumping to using a with-statement.
