### 2019-02-25 d9f6e39
Changed version to signify end of v1.1.1 development cycle.
