### 2019-02-25 2a2f7b1
Moved the _dummyMPI module to e13tools.dummyMPI. Moved several non-PRISM-specific functions from _internal to e13tools.utils. Updated requirement for e13Tools.
