+++ prism/tests/test_internal.py
@@ -5,6 +5,7 @@
 import logging
 from os import path
 from sys import platform
+import warnings
 
 # Package imports
 from e13tools.core import InputError
@@ -14,12 +15,9 @@ import pytest
 # PRISM imports
 from prism.__version__ import compat_version, prism_version
 from prism._internal import (FeatureWarning, PRISM_Logger, RequestError,
-                             RequestWarning, docstring_append, docstring_copy,
-                             docstring_substitute, check_instance, check_vals,
-                             check_compatibility, convert_str_seq, delist,
+                             RequestWarning, check_vals, check_compatibility,
                              get_PRISM_File, get_info, getCLogger, getRLogger,
-                             import_cmaps, np_array, raise_error,
-                             raise_warning, rprint)
+                             np_array)
 
 # Save the path to this directory
 dirpath = path.dirname(__file__)
@@ -29,6 +27,12 @@ win32 = platform.startswith('win')
 
 
 # %% PYTEST CLASSES AND FUNCTIONS
+# Pytest for FeatureWarning
+def test_FeatureWarning():
+    with pytest.warns(FeatureWarning):
+        warnings.warn("WARNING", FeatureWarning)
+
+
 # Pytest for PRISM_File class
 class Test_PRISM_File(object):
     def test(self, tmpdir):
@@ -48,71 +52,6 @@ class Test_PRISM_File(object):
         assert path.exists(filename)
 
 
-# Pytest for the custom function decorators
-class TestDecorators(object):
-    # Create method with no docstring that is appended
-    @docstring_append("appended")
-    def append_method1(self):
-        pass
-
-    # Create method with a docstring that is appended
-    @docstring_append("appended")
-    def append_method2(self):
-        """original """
-
-    # Check if docstring_append works correctly
-    def test_docstring_append(self):
-        assert self.append_method1.__doc__ == "appended"
-        assert self.append_method2.__doc__ == "original appended"
-
-    # Create method with no docstring at all
-    def empty_method(self):
-        pass
-
-    # Create method that copies an empty docstring
-    @docstring_copy(empty_method)
-    def copy_method1(self):
-        pass
-
-    # Create method that copies a docstring
-    @docstring_copy(append_method1)
-    def copy_method2(self):
-        pass
-
-    # Check if docstring_copy works correctly
-    def test_docstring_copy(self):
-        assert self.copy_method1.__doc__ is None
-        assert self.copy_method1.__doc__ == self.empty_method.__doc__
-        assert self.copy_method2.__doc__ == self.append_method1.__doc__
-
-    # Check if providing both args and kwargs raises an error
-    with pytest.raises(InputError):
-        @docstring_substitute("positional", x="keyword")
-        def substitute_method1(self):
-            pass
-
-    # Create method using args substitutes
-    @docstring_substitute("positional")
-    def substitute_method2(self):
-        """%s"""
-
-    # Create method using kwargs substitutes
-    @docstring_substitute(x="keyword")
-    def substitute_method3(self):
-        """%(x)s"""
-
-    # Check if providing args to a method with no docstring raises an error
-    with pytest.raises(InputError):
-        @docstring_substitute("positional")
-        def substitute_method4(self):
-            pass
-
-    # Check if docstring_substitute works correctly
-    def test_docstring_substitute(self):
-        assert self.substitute_method2.__doc__ == "positional"
-        assert self.substitute_method3.__doc__ == "keyword"
-
-
 # Pytest for the check_compatibility function
 def test_check_compatibility():
     # Check if current version is compatible (it should)
[…]
