### 2019-02-26 c492dad
Added missing documentation on the new (experimental) backup system.
