+++ prism/utils/mcmc.py
@@ -77,6 +77,7 @@ def get_lnpost_fn(ext_lnpost, pipeline_obj, *, emul_i=None, unit_space=True,
     --------
     :func:`~get_walkers`: Analyzes proposed `init_walkers` and returns valid \
         `p0_walkers`.
+
     :attr:`~prism.Pipeline.worker_mode`: Special context manager within which \
         all code is executed in worker mode.
 
@@ -244,6 +245,7 @@ def get_walkers(pipeline_obj, *, emul_i=None, init_walkers=None,
     --------
     :func:`~get_lnpost_fn`: Returns a function definition \
         ``get_lnpost(par_set, *args, **kwargs)``.
+
     :attr:`~prism.Pipeline.worker_mode`: Special context manager within which \
         all code is executed in worker mode.
 
