### 2019-02-27 b98c75b
Cleaned up the code a bit by avoiding catching AttributeErrors. Checking if Coveralls can actually report coverage properly.
