### 2019-02-27 0b5a913
Switched to using Coveralls now, as it actually does report coverage properly (unlike CodeCov).
