+++ README.rst
@@ -156,6 +156,6 @@ And, finally, if you use *PRISM* as part of your workflow in a scientific public
 .. |ReadTheDocs| image:: https://img.shields.io/readthedocs/prism-tool/latest.svg?logo=read%20the%20docs&logoColor=white&label=Docs
     :target: https://prism-tool.readthedocs.io/en/latest
     :alt: ReadTheDocs - Build Status
-.. |Coverage| image:: https://img.shields.io/codecov/c/github/1313e/PRISM/master.svg?logo=codecov&logoColor=white&label=Coverage
-    :target: https://codecov.io/gh/1313e/PRISM/branches/master
-    :alt: CodeCov - Coverage Status
+.. |Coverage| image:: https://img.shields.io/coveralls/github/1313e/PRISM/master.svg?logo=reverbnation&logoColor=white&label=Coverage
+    :target: https://coveralls.io/github/1313e/PRISM?branch=master
+    :alt: Coveralls - Coverage Status
