+++ .codecov.yml
@@ -1,17 +1,0 @@
-codecov:
-    ci:
-        - !appveyor
-    notify:
-        require_ci_to_pass: yes
-  
-
-coverage:
-    range: 0..100
-    round: down
-    precision: 2
-    status:
-        project:
-            default:
-              	target: 95%
-        patch: false
-comment: off
