### 2019-02-27 4250fe9
Move v1.0.x branch to Coveralls as well.
