### 2019-02-28 a8c5454
Split the __set_model_data() and __set_model_parameters() methods up into a private and public part. Added two new modellink utility functions called convert_data() and convert_parameters() (the public parts). These functions take an input and convert it into a full dict to be used by ModelLink, taking into account all versatility formatting. Added a few notes in the documentation that these functions can be used to check if one has formatted their parameters/data properly. Changed some sentences in the docstrings of the ModelLink class to better describe the formatting of the parameters/data 
[…]
