+++ prism/modellink/_modellink.py
@@ -32,7 +32,7 @@ from prism._internal import (FeatureWarning, PRISM_Comm, RequestWarning,
                              check_vals, getCLogger, np_array)
 
 # All declaration
-__all__ = ['ModelLink', 'test_subclass']
+__all__ = ['ModelLink', 'convert_data', 'convert_parameters', 'test_subclass']
 
 
 # %% MODELLINK CLASS DEFINITION
@@ -114,7 +114,8 @@ class ModelLink(object, metaclass=abc.ABCMeta):
         required to provide an estimate for every parameter. The estimates are
         used to draw illustrative lines when making projection figures.
         An example of a model parameters file can be found in the 'data' folder
-        of the *PRISM* package.
+        of the *PRISM* package. If required, one can use the
+        :func:`~check_parameters` function to validate their formatting.
 
         Formatting :
             ``{par_name: [lower_bnd, upper_bnd, par_est]}``
@@ -150,7 +151,8 @@ class ModelLink(object, metaclass=abc.ABCMeta):
         :math:`f(x)`.
 
         An example of a model data file can be found in the 'data' folder of
-        the *PRISM* package.
+        the *PRISM* package. If required, one can use the :func:`~check_data`
+        function to validate their formatting.
 
         Formatting :
             ``{(data_idx_0, data_idx_1, ..., data_idx_n): [data_val,`` \
@@ -827,41 +829,12 @@ class ModelLink(object, metaclass=abc.ABCMeta):
         """
 
         # Obtain default model parameters
-        model_parameters = sdict(self.get_default_model_parameters())
+        model_parameters =\
+            convert_parameters(self.get_default_model_parameters())
 
-        # If no additional model parameters information is given
-        if add_model_parameters is None:
-            pass
-
-        # If a parameter file is given
-        elif isinstance(add_model_parameters, str):
-            # Obtain absolute path to given file
-            par_file = path.abspath(add_model_parameters)
-
-            # Read the parameter file in as a string
-            pars = np.genfromtxt(par_file, dtype=(str), delimiter=':',
-                                 autostrip=True)
-
-            # Make sure that pars is 2D
-            pars = np_array(pars, ndmin=2)
-
-            # Combine default parameters with read-in parameters
-            model_parameters.update(pars)
-
-        # If a parameter dict is given
-        elif isinstance(add_model_parameters, dict):
-            model_parameters.update(add_model_parameters)
-
-        # If anything else is given
-        else:
-            # Check if it can be converted to a dict
-            try:
-                par_dict = sdict(add_model_parameters)
-            except Exception:
-                raise TypeError("Input model parameters cannot be converted to"
-                                " type 'dict'!")
-            else:
-                model_parameters.update(par_dict)
+        # If additional model parameters information is given, add it
+        if add_model_parameters is not None:
+            model_parameters.update(convert_parameters(add_model_parameters))
 
         # Save number of model parameters
         n_par = len(model_parameters.keys())
@@ -877,36 +850,11 @@ class ModelLink(object, metaclass=abc.ABCMeta):
         self._par_est = []
 
         # Save model parameters as class properties
-        for i, (name, values_str) in enumerate(model_parameters.items()):
-            # Convert values_str to values
-            values = convert_str_seq(values_str)
-
-            # Save parameter name and range
-            self._par_name.append(check_vals(name, 'par_name', 'str'))
-            self._par_rng[i] = check_vals(values[:2], 'par_rng[%s]' % (name),
-                                          'float')
-
-            # Check if a float parameter estimate was provided
-            try:
-                est = check_vals(values[2], 'par_est[%s]' % (name), 'float')
-            # If no estimate was provided, save it as None
-            except IndexError:
-                self._par_est.append(None)
-            # If no float was provided, check if it was None
[…]
