### 2019-02-28 7e9dba6
Changed the terminology list to a glossary, such that specific terms are searchable. Made some more optimizations concerning lay-outs for the docs.
