### 2019-03-01 e56b517
Moved the utility functions in _modellink to their own proper submodule (modellink.utils). Fixed the Pipeline.__init__ docstring to properly mention that the dummyMPI module is now located in the e13Tools package.
