+++ prism/modellink/__init__.py
@@ -12,10 +12,11 @@ and various default/example :class:`~ModelLink` subclasses.
 # %% IMPORTS
 # Import modellink modules
 from ._gaussian_link import GaussianLink
-from . import _modellink
-from ._modellink import *
+from ._modellink import ModelLink
 from ._sine_wave_link import SineWaveLink
+from . import utils
+from .utils import *
 
 # All declaration
-__all__ = ['GaussianLink', 'SineWaveLink']
-__all__.extend(_modellink.__all__)
+__all__ = ['GaussianLink', 'ModelLink', 'SineWaveLink', 'utils']
+__all__.extend(utils.__all__)
