+++ prism/modellink/tests/test_modellink.py
@@ -11,7 +11,7 @@ import pytest
 
 # PRISM imports
 from prism._internal import RequestWarning, MPI
-from prism.modellink import ModelLink, test_subclass as _test_subclass
+from prism.modellink import ModelLink
 from prism.modellink.tests.modellink import GaussianLink2D, GaussianLink3D
 
 
@@ -30,90 +30,6 @@ model_parameters_outside_est = path.join(dirpath,
 
 
 # %% CUSTOM CLASSES
-# Custom invalid ModelLink class
-class InvalidModelLink(ModelLink):
-    def __init__(self, *args, **kwargs):
-        pass
-
-
-# Custom improper ModelLink class
-class ImproperModelLink(ModelLink):
-    def __init__(self, *args, **kwargs):
-        pass
-
-    def call_model(self, *args, **kwargs):
-        super().call_model(*args, **kwargs)
-
-    def get_md_var(self, *args, **kwargs):
-        super().get_md_var(*args, **kwargs)
-
-
-# Custom ModelLink class with no call_model
-class NoCallModelLink(ModelLink):
-    def __init__(self, *args, **kwargs):
-        self.MPI_call = True
-        super().__init__(*args, **kwargs)
-
-    def call_model(self, emul_i, par_set, data_idx):
-        super().call_model(emul_i, par_set, data_idx)
-
-    def get_md_var(self, emul_i, par_set, data_idx):
-        super().get_md_var(emul_i, par_set, data_idx)
-
-
-# Custom ModelLink class that does not accept the correct call_model arguments
-class WrongCallModelLink(ModelLink):
-    def __init__(self, *args, **kwargs):
-        self.MPI_call = True
-        super().__init__(*args, **kwargs)
-
-    def call_model(self, emul_i):
-        pass
-
-    def get_md_var(self, *args, **kwargs):
-        super().get_md_var(*args, **kwargs)
-
-
-# Custom ModelLink class that accepts too many call_model arguments
-class ManyCallModelLink(ModelLink):
-    def __init__(self, *args, **kwargs):
-        self.MPI_call = True
-        super().__init__(*args, **kwargs)
-
-    def call_model(self, emul_i, par_set, data_idx, test):
-        pass
-
-    def get_md_var(self, *args, **kwargs):
-        super().get_md_var(*args, **kwargs)
-
-
-# Custom ModelLink class with no get_md_var()
-class NoMdVarModelLink(ModelLink):
-    def call_model(self, emul_i, par_set, data_idx):
-        return([1]*len(data_idx))
-
-    def get_md_var(self, emul_i, par_set, data_idx):
-        super().get_md_var(emul_i, par_set, data_idx)
-
-
-# Custom ModelLink class that does not accept the correct get_md_var arguments
-class WrongMdVarModelLink(ModelLink):
-    def call_model(self, emul_i, par_set, data_idx):
-        return([1]*len(data_idx))
-
-    def get_md_var(self, emul_i):
-        pass
-
-
-# Custom ModelLink class that accepts too many get_md_var arguments
-class ManyMdVarModelLink(ModelLink):
-    def call_model(self, emul_i, par_set, data_idx):
-        return([1]*len(data_idx))
-
-    def get_md_var(self, emul_i, par_set, data_idx, test):
-        pass
-
-
 # Custom ModelLink class that uses the backup system
 class BackupModelLink(ModelLink):
     def call_model(self, data_idx, *args, **kwargs):
@@ -305,68 +221,3 @@ class Test_backup_system(object):
         # Make sure no backup file was created
         with pytest.raises(OSError):
             modellink_obj._read_backup(emul_i)
-
-
-# Pytest for test_subclass function
-class Test_test_subclass(object):
[…]
