+++ prism/modellink/utils.py
@@ -1,0 +1,422 @@
+# -*- coding: utf-8 -*-
+
+"""
+ModelLink Utilities
+===================
+Provides several utility functions concerning the
+:class:`~prism.modellink.ModelLink` class.
+
+"""
+
+
+# %% IMPORTS
+# Built-in imports
+from inspect import _VAR_KEYWORD, _VAR_POSITIONAL, _empty, isclass, signature
+from os import path
+import warnings
+
+# Package imports
+from e13tools import InputError
+from e13tools.utils import check_instance, convert_str_seq
+import numpy as np
+from numpy.random import rand
+from sortedcontainers import SortedDict as sdict
+
+# PRISM imports
+from prism._internal import PRISM_Comm, RequestWarning, check_vals, np_array
+
+# All declaration
+__all__ = ['convert_data', 'convert_parameters', 'test_subclass']
+
+
+# %% UTILITY FUNCTIONS
+# This function converts provided model data into format used by PRISM
+def convert_data(model_data):
+    """
+    Converts the provided `model_data` into a full data dict, taking into
+    account all formatting options, and returns it.
+
+    This function can be used externally to check how the provided `model_data`
+    would be interpreted when provided to the
+    :class:`~prism.modellink.ModelLink` subclass. Its output can be used for
+    the 'model_data' input argument.
+
+    Parameters
+    ----------
+    model_data : array_like, dict or str
+        Anything that can be converted to a dict that provides model data
+        information.
+
+    Returns
+    -------
+    data_dict : dict
+        Dict with the provided `model_data` converted to its full format.
+
+    """
+
+    # If a data file is given
+    if isinstance(model_data, str):
+        # Obtain absolute path to given file
+        data_file = path.abspath(model_data)
+
+        # Read the data file in as a string
+        data_points = np.genfromtxt(data_file, dtype=(str), delimiter=':',
+                                    autostrip=True)
+
+        # Make sure that data_points is 2D
+        data_points = np_array(data_points, ndmin=2)
+
+        # Convert read-in data to dict
+        model_data = dict(data_points)
+
+    # If a data dict is given
+    elif isinstance(model_data, dict):
+        model_data = dict(model_data)
+
+    # If anything else is given
+    else:
+        # Check if it can be converted to a dict
+        try:
+            model_data = dict(model_data)
+        except Exception:
+            raise TypeError("Input model data cannot be converted to type "
+                            "'dict'!")
+
+    # Make empty data_dict
+    data_dict = dict()
+
+    # Loop over all items in model_data
+    for key, value in model_data.items():
+        # Convert key to an actual data_idx
+        idx = convert_str_seq(key)
+
+        # Check if tmp_idx is not empty
+        if not len(idx):
+            raise InputError("Model data contains a data point with no "
+                             "identifier!")
+
+        # Convert value to an actual data point
+        data = convert_str_seq(value)
+
+        # Check if provided data value is correct
+        val = check_vals(data[0], 'data_val%s' % (idx), 'float')
+
+        # Extract data error and space
+        # If length is two, centered error and no data space were given
+        if(len(data) == 2):
+            err = [check_vals(data[1], 'data_err%s' % (idx), 'float', 'pos')]*2
+            spc = 'lin'
+
+        # If length is three, there are two possibilities
+        elif(len(data) == 3):
+            # If the third column contains a string, it is the data space
+            if isinstance(data[2], str):
+                err = [check_vals(data[1], 'data_err%s' % (idx),
+                                  'float', 'pos')]*2
+                spc = data[2]
+
+            # If the third column contains no string, it is error interval
+            else:
+                err = check_vals(data[1:3], 'data_err%s' % (idx),
[…]

[…]
