### 2019-03-04 b85c7e3
Setting the name, call_type and MPI_call properties of a ModelLink subclass outside of its constructor, now raises a warning (since it may cause unexpected effects). Fixed a bug where the MPI_call property returned the wrong internal property value.
