+++ prism/modellink/tests/test_modellink.py
@@ -6,6 +6,7 @@ from os import path
 
 # Package imports
 from e13tools.core import InputError
+from e13tools.utils import get_outer_frame
 import numpy as np
 import pytest
 
@@ -69,9 +70,6 @@ class Test_ModelLink_Exceptions(object):
     # Also include an invalid call type
     def test_invalid_call_type(self):
         modellink_obj = GaussianLink2D()
-        modellink_obj.call_type = 'single'
-        modellink_obj.call_type = 'multi'
-        modellink_obj.call_type = 'hybrid'
         with pytest.raises(ValueError):
             modellink_obj.call_type = 'invalid'
 
@@ -111,23 +109,23 @@ class Test_ModelLink_Versatility(object):
     def test_ext_mod_par_dict(self):
         par_dict = {'A': [1, 5, None],
                     'B': [1, 3]}
-        model_link = GaussianLink2D(model_parameters=par_dict)
-        assert model_link._par_est == [None, None]
-        repr(model_link)
+        modellink_obj = GaussianLink2D(model_parameters=par_dict)
+        assert modellink_obj._par_est == [None, None]
+        repr(modellink_obj)
 
     # Create a GaussianLink2D object with externally defined mod_par list
     def test_ext_mod_par_list(self):
         par_list = [['A', [1, 5, 2.5]],
                     ['B', [1, 3, 2]]]
-        model_link = GaussianLink2D(model_parameters=par_list)
-        repr(model_link)
+        modellink_obj = GaussianLink2D(model_parameters=par_list)
+        repr(modellink_obj)
 
     # Create a GaussianLink2D object and test the value conversions
     def test_convert_to_space(self):
-        model_link = GaussianLink2D()
-        assert np.isclose(model_link._to_par_space([0.2, 0.7]),
+        modellink_obj = GaussianLink2D()
+        assert np.isclose(modellink_obj._to_par_space([0.2, 0.7]),
                           [1.8, 2.4]).all()
-        assert np.isclose(model_link._to_unit_space([1.8, 2.4]),
+        assert np.isclose(modellink_obj._to_unit_space([1.8, 2.4]),
                           [0.2, 0.7]).all()
 
     # Create a GaussianLink3D object with externally defined mod_data dict
@@ -135,31 +133,54 @@ class Test_ModelLink_Versatility(object):
         data_dict = {(1, 2): [1, 0.05],
                      3: [2, 0.05],
                      4: [3, 0.05]}
-        model_link = GaussianLink3D(model_parameters=model_parameters_3D,
-                                    model_data=data_dict)
-        repr(model_link)
+        modellink_obj = GaussianLink3D(model_parameters=model_parameters_3D,
+                                       model_data=data_dict)
+        repr(modellink_obj)
 
     # Create a GaussianLink3D object with externally defined mod_data list
     def test_ext_mod_data_list(self):
         data_list = [[(1, 2), [1, 0.05]],
                      [3, [2, 0.05]],
                      [4, [3, 0.05]]]
-        model_link = GaussianLink3D(model_parameters=model_parameters_3D,
-                                    model_data=data_list)
-        repr(model_link)
+        modellink_obj = GaussianLink3D(model_parameters=model_parameters_3D,
+                                       model_data=data_list)
+        repr(modellink_obj)
 
     # Create a GaussianLink3D object with mod_data file with double errors
     def test_double_errors(self):
-        model_link = GaussianLink3D(model_parameters=model_parameters_3D,
-                                    model_data=model_data_double)
-        repr(model_link)
+        modellink_obj = GaussianLink3D(model_parameters=model_parameters_3D,
+                                       model_data=model_data_double)
+        repr(modellink_obj)
 
     # Create a GaussianLink3D object with different data_idx and data_spc
     # Also include a data_idx sequence
[…]
