### 2019-03-04 7001bb0
Fixed a bug where the MPI_call property returned the wrong internal property value.
