### 2019-03-04 cfc8137
New working directories can now be requested by providing a True to the Pipeline class. Providing integers for this is no longer allowed.
