### 2019-03-04 f682af9
Improved the backup system to automatically create a random suffix for every backup file created, such that older ones can be retained. This suffix can now be provided to the _read_backup()-method as well, to read a specific backup instead of the last one created.
