### 2019-03-04 fb9aa5f
(Hopefully) Fixed a bug in the pytests for the backup system where the last suffix was extracted incorrectly.
