### 2019-03-05 84c325b
Fixed a bug where a pytest would fail if the suffix of a backup file ended with any character in 'hdf5'.
