+++ prism/modellink/tests/test_modellink.py
@@ -218,9 +218,7 @@ class Test_backup_system(object):
             assert data['kwargs'] == {'mod_set': mod_set}
 
             # Try loading the backup again, this time using the suffix
-            print(filename)
-            suffix = filename.partition(
-                '%s(' % (modellink_obj._name))[2].rstrip(').hdf5')
+            suffix = filename.partition('%s(' % (modellink_obj._name))[2][:-6]
             filename, data = modellink_obj._read_backup(emul_i, suffix=suffix)
 
             # Check that all data is correct
