### 2019-03-07 d75e10a
The assigned data_idx to every core are now also saved on the controller.
