### 2019-03-07 d2ec03c
Finally try to get CodeCov working properly.
