+++ .codecov.yml
@@ -1,0 +1,13 @@
+codecov:
+    notify:
+        require_ci_to_pass: yes
+  
+
+coverage:
+    range: 0..100
+    round: down
+    precision: 2
+    status:
+        project: yes
+        patch: no
+comment: off
