### 2019-03-07 23a5433
I might have found the problem with CodeCov not reporting coverage properly.
