+++ .codecov.yml
@@ -1,13 +1,15 @@
 codecov:
     notify:
         require_ci_to_pass: yes
-  
 
 coverage:
     range: 0..100
     round: down
     precision: 2
     status:
-        project: yes
+        project:
+            default:
+                target: 99%
         patch: no
+        changes: yes
 comment: off
