+++ .codecov.yml
@@ -3,13 +3,17 @@ codecov:
         require_ci_to_pass: yes
 
 coverage:
-    range: 0..100
+    range: 90..100
     round: down
     precision: 2
     status:
         project:
             default:
                 target: 99%
-        patch: no
+        patch:
+            default:
+                target: 95%
         changes: yes
-comment: off
+
+comment:
+    require_changes: yes
