### 2019-03-07 e4e49c0
Back to CodeCov now that I finally figured out what the problem was.
