+++ README.rst
@@ -164,6 +164,6 @@ And, finally, if you use *PRISM* as part of your workflow in a scientific public
 .. |ReadTheDocs| image:: https://img.shields.io/readthedocs/prism-tool/v1.0.x.svg?logo=read%20the%20docs&logoColor=white&label=Docs
     :target: https://prism-tool.readthedocs.io/en/v1.0.x
     :alt: ReadTheDocs - Build Status
-.. |Coverage| image:: https://img.shields.io/coveralls/github/1313e/PRISM/v1.0.x.svg?logo=reverbnation&logoColor=white&label=Coverage
-    :target: https://coveralls.io/github/1313e/PRISM?branch=v1.0.x
-    :alt: Coveralls - Coverage Status
+.. |Coverage| image:: https://img.shields.io/codecov/c/github/1313e/PRISM/v1.0.x.svg?logo=codecov&logoColor=white&label=Coverage
+    :target: https://codecov.io/gh/1313e/PRISM/branches/v1.0.x
+    :alt: CodeCov - Coverage Status
