+++ setup.cfg
@@ -7,7 +7,7 @@ testpaths = ./prism
 
 [coverage:run]
 parallel = true
-source = ./prism
+include = prism/*
 omit =
   prism/tests/*
   prism/*/tests/*
