+++ .codecov.yml
@@ -1,0 +1,19 @@
+codecov:
+    notify:
+        require_ci_to_pass: yes
+
+coverage:
+    range: 90..100
+    round: down
+    precision: 2
+    status:
+        project:
+            default:
+                target: 99%
+        patch:
+            default:
+                target: 95%
+        changes: yes
+
+comment:
+    require_changes: yes
