### 2019-03-08 4ec9b9a
The convert_parameters()-function now additionally checks if a provided parameter range is valid (lower bound is below the upper bound).
