+++ prism/modellink/tests/test_modellink.py
@@ -25,6 +25,8 @@ model_data_single = path.join(dirpath, 'data/data_gaussian_single.txt')
 model_data_double = path.join(dirpath, 'data/data_gaussian_double.txt')
 model_data_types = path.join(dirpath, 'data/data_gaussian_types.txt')
 model_parameters_3D = path.join(dirpath, 'data/parameters_gaussian_3D.txt')
+model_parameters_invalid_rng = path.join(dirpath,
+                                         'data/parameters_invalid_rng.txt')
 model_parameters_invalid_est = path.join(dirpath,
                                          'data/parameters_invalid_est.txt')
 model_parameters_outside_est = path.join(dirpath,
@@ -56,6 +58,11 @@ class Test_ModelLink_Exceptions(object):
     def test_invalid_model_par(self):
         with pytest.raises(TypeError):
             GaussianLink2D(model_parameters=np.array([1]))
+
+    # Try to create a GaussianLink2D object with invalid parameter ranges
+    def test_invalid_model_par_rng(self):
+        with pytest.raises(ValueError):
+            GaussianLink2D(model_parameters=model_parameters_invalid_rng)
 
     # Try to create a GaussianLink2D object with invalid parameter estimates
     def test_invalid_model_par_est(self):
