+++ prism/modellink/__init__.py
@@ -10,10 +10,14 @@ and various default/example :class:`~ModelLink` subclasses.
 
 
 # %% IMPORTS
-# Import modellink modules
-from ._gaussian_link import GaussianLink
+# Import ModelLink abstract base class
 from ._modellink import ModelLink
+
+# Import ModelLink subclasses
+from ._gaussian_link import GaussianLink
 from ._sine_wave_link import SineWaveLink
+
+# Import modellink modules
 from . import utils
 from .utils import *
 
