### 2019-03-08 93b53d1
Moved the Emulator class to its own subpackage 'emulator'. This will allow future specialized Emulator classes to be added more easily. Made several changes to imports to reflect this change.
