### 2019-03-08 e69e7da
Change CodeCov rounding to nearest. Hopefully the range will now be shown correctly.
