+++ .codecov.yml
@@ -3,8 +3,8 @@ codecov:
         require_ci_to_pass: yes
 
 coverage:
-    range: 90..100
-    round: down
+    range: 90%...100%
+    round: nearest
     precision: 2
     status:
         project:
