### 2019-03-12 df81c4b
Fixed a bug where the root logger would still handle logging messages of PRISM, as long as it has a handler whose loggingLevel is low enough. Now, the PRISM root logger no longer propagates logging messages to any of its ancestors.
