### 2019-03-12 c3f9500
Changed docstring of the evaluate()-method of the Pipeline class to be more clear about when it returns and when it prints.
