### 2019-03-13 7bb7e47
Small improvements to the online documentation. 2D projections now use the same axis labels as 3D projections.
