### 2019-03-13 8e405d7
Small improvements to the online documentation. 2D projections now use the same axis labels as 3D projections.
