### 2019-03-13 cc91fed
Added property setters to the Pipeline and Projection classes, except for the implausibility properties. The PRISM parameters file is now only read once, when the Pipeline class is initialized. Added a new Pipeline property 'prism_dict', which stores the dict of read-in PRISM parameters. Whenever a new emulator is made, it will take its parameters from 'prism_dict'.
