+++ prism/_projection.py
@@ -34,7 +34,7 @@ from tqdm import tqdm
 # PRISM imports
 from prism._docstrings import (def_par_doc, draw_proj_fig_doc, get_emul_i_doc,
                                hcube_doc, proj_data_doc, proj_par_doc_d,
-                               proj_par_doc_s, read_par_doc, save_data_doc_pr,
+                               proj_par_doc_s, save_data_doc_pr, set_par_doc,
                                user_emul_i_doc)
 from prism._internal import (RequestError, RequestWarning, check_vals,
                              getCLogger, np_array)
@@ -241,7 +241,7 @@ class Projection(object):
                         self.__get_proj_data(hcube)
                 # Otherwise, the used resolution is the current resolution
                 else:
-                    proj_res = self.__res
+                    proj_res = self._res
 
                 # Draw projection figure
                 if(len(hcube) == 1):
@@ -276,24 +276,32 @@ class Projection(object):
     @property
     def proj_res(self):
         """
-        int: Number of emulator evaluations used to generate the grid for the
-        last created projection figures.
+        int: Number of emulator evaluations that will be used to generate the
+        grid for the projection figures.
 
         """
 
-        return(getattr(self, '__res', None))
+        return(getattr(self, '_res', None))
+
+    @proj_res.setter
+    def proj_res(self, proj_res):
+        self._res = check_vals(proj_res, 'proj_res', 'int', 'pos')
 
     @property
     def proj_depth(self):
         """
-        int: Number of emulator evaluations used to generate the samples in
-        every grid point for the last created projection figures.
-        Note that when making 2D projections of nD models, the used depth was
-        this number multiplied by :attr:`~proj_res`.
+        int: Number of emulator evaluations that will be used to generate the
+        samples in every grid point for the projection figures.
+        Note that when making 2D projections of nD models, the used depth will
+        be this number multiplied by :attr:`~proj_res`.
 
         """
 
-        return(getattr(self, '__depth', None))
+        return(getattr(self, '_depth', None))
+
+    @proj_depth.setter
+    def proj_depth(self, proj_depth):
+        self._depth = check_vals(proj_depth, 'proj_depth', 'int', 'pos')
 
     # %% HIDDEN CLASS METHODS
     # This function draws the 2D projection figure
@@ -871,7 +879,7 @@ class Projection(object):
                     'proj_depth': '250'}
 
         # Return it
-        return(par_dict)
+        return(sdict(par_dict))
 
     # This function returns default projection input arguments
     def __get_default_input_arguments(self):
@@ -914,41 +922,33 @@ class Projection(object):
                        'figsize_r': figsize_r}
 
         # Return it
-        return(kwargs_dict)
+        return(sdict(kwargs_dict))
 
-    # Read in the parameters from the provided parameter file
-    @docstring_append(read_par_doc.format("Projection"))
-    def __read_parameters(self):
-        # Log that the PRISM parameter file is being read
+    # Set the parameters that were read in from the provided parameter file
+    @docstring_append(set_par_doc.format("Projection"))
+    def __set_parameters(self):
+        # Log that the projection parameters are being set
         logger = getCLogger('INIT')
-        logger.info("Reading projection parameters.")
+        logger.info("Setting projection parameters.")
 
         # Obtaining default projection parameter dict
         par_dict = self.__get_default_parameters()
 
-        # Read in data from provided PRISM parameters file
-        if self._prism_file is not None:
-            pipe_par = np.genfromtxt(self._prism_file, dtype=(str),
-                                     delimiter=':', autostrip=True)
-
-            # Make sure that pipe_par is 2D
-            pipe_par = np_array(pipe_par, ndmin=2)
-
[…]

[…]
