### 2019-03-13 95647db
Removed the percentage signs from the .codecov.yml file.
