+++ .codecov.yml
@@ -3,7 +3,7 @@ codecov:
         require_ci_to_pass: yes
 
 coverage:
-    range: 90%...100%
+    range: 90..100
     round: nearest
     precision: 2
     status:
