### 2019-03-13 c685988
Changed the getLogger()-function such to only apply filters if the requested PRISM_Logger instance did not already exist.
