### 2019-03-14 99ce0fb
Added property setters for the implausibility parameters. An 'impl_par' list can now be provided to the analyze()-method of the Pipeline class to change the used implausibility parameters without requiring to reload the entire pipeline. Updated documentation in several places to indicate that Pipeline properties can now be set as well. The check_vals()-function can now also check for complex numbers.
