+++ prism/_pipeline.py
@@ -26,8 +26,8 @@ from e13tools import InputError, ShapeError
 from e13tools.math import nCr
 from e13tools.sampling import lhd
 from e13tools.utils import (convert_str_seq, delist, docstring_append,
-                            docstring_copy, docstring_substitute, raise_error,
-                            raise_warning)
+                            docstring_copy, docstring_substitute,
+                            get_outer_frame, raise_error, raise_warning)
 import numpy as np
 from numpy.random import normal, random
 from sortedcontainers import SortedDict as sdict
@@ -37,9 +37,9 @@ from prism._docstrings import (call_emul_i_doc, call_model_doc_s,
                                call_model_doc_m, def_par_doc, emul_s_seq_doc,
                                ext_mod_set_doc, ext_real_set_doc_d,
                                ext_real_set_doc_s, ext_sam_set_doc,
-                               impl_cut_doc, impl_temp_doc, paths_doc_d,
-                               paths_doc_s, save_data_doc_p, set_par_doc,
-                               std_emul_i_doc, user_emul_i_doc)
+                               impl_cut_doc, paths_doc_d, paths_doc_s,
+                               save_data_doc_p, set_par_doc, std_emul_i_doc,
+                               user_emul_i_doc)
 from prism._internal import (PRISM_Comm, RequestError, RequestWarning,
                              check_vals, getCLogger, get_PRISM_File,
                              getRLogger, move_logger, np_array,
@@ -593,12 +593,101 @@ class Pipeline(Projection, object):
     @property
     def impl_cut(self):
         """
-        list of int: The non-wildcard univariate implausibility cut-off values
-        for an emulator iteration.
+        list of float: The non-wildcard univariate implausibility cut-off
+        values for an emulator iteration. Setting it with the reduced
+        implausibility cut-off list will change the values of :attr:`~cut_idx`
+        and this property at the last emulator iteration.
 
         """
 
         return(self._impl_cut)
+
+    @impl_cut.setter
+    def impl_cut(self, impl_cut):
+        # Only the controller can set this value
+        if self._is_controller:
+            # Make logger
+            logger = getRLogger('CHECK')
+
+            # Get last emul_i
+            emul_i = self._emulator._emul_i
+
+            # Raise error if emul_i is 0
+            if not emul_i:
+                err_msg = "Emulator is not built yet!"
+                raise_error(err_msg, RequestError, logger)
+
+            # Do some logging
+            logger.info("Checking compatibility of provided implausibility "
+                        "cut-off values.")
+
+            # Check if analyze() is calling this setter
+            caller_frame = get_outer_frame(self.analyze)
+
+            # Check if the last iteration has already been analyzed
+            try:
+                if self._n_eval_sam[emul_i]:
+                    # If so, check if this setter was called by analyze()
+                    if caller_frame is None:
+                        # If not, raise error
+                        err_msg = ("Pipeline property 'impl_cut' can only be "
+                                   "set if the current emulator iteration has "
+                                   "not been analyzed yet!")
+                        raise_error(err_msg, RequestError, logger)
+
+            # If n_eval_sam at emul_i does not exist, no analyze has been done
+            except IndexError:
+                pass
+
+            # Complete the impl_cut list
+            # Obtain the first impl_cut value
+            try:
+                impl_cut[0] = check_vals(impl_cut[0], 'impl_cut[0]', 'float',
+                                         'nneg')
+            except IndexError:
+                err_msg = ("Provided implausibility cut-off list contains no "
+                           "elements!")
+                raise_error(err_msg, ValueError, logger)
+
[…]
