### 2019-03-15 09a9fd4
The *PRISM* parameters can now be provided as a dict, str or anything that can be converted to a dict. Fixed some parts of the online documentation not showing properly. Renamed the 'prism_file' input argument to 'prism_par'. Added a FutureWarning when somebody attempts to use 'prism_file'. The check_vals()-function now additionally checks if the provided values is/are not empty. Rewritten the representation of a Pipeline object to use a dict-format for the 'prism_par' argument. Fixed a bug where providing the criterion or pot_active_par as anything besides a string would not work (as it assum
[…]
