+++ prism/tests/test_pipeline_emulator_projection.py
@@ -133,7 +133,7 @@ class Test_Pipeline_Gaussian2D(object):
         working_dir = path.basename(tmpdir.strpath)
         model_link = GaussianLink2D()
         return(Pipeline(model_link, root_dir=root_dir, working_dir=working_dir,
-                        prism_file=prism_file_default))
+                        prism_par=prism_file_default))
 
     # Check if representation can be called
     def test_repr(self, pipe):
@@ -265,7 +265,7 @@ class Test_Pipeline_Gaussian3D(object):
         model_link = GaussianLink3D(model_parameters=model_parameters_3D,
                                     model_data=model_data_single)
         return(Pipeline(model_link, root_dir=root_dir, working_dir=working_dir,
-                        prism_file=prism_file_default))
+                        prism_par=prism_file_default))
 
     # Check if representation can be called
     def test_repr(self, pipe):
@@ -325,7 +325,7 @@ class Test_Pipeline_Gaussian3D_1_data(object):
         model_link = GaussianLink3D(model_parameters=model_parameters_3D,
                                     model_data={2: [2, 0.05]})
         return(Pipeline(model_link, root_dir=root_dir, working_dir=working_dir,
-                        prism_file=prism_file_default))
+                        prism_par=prism_file_default))
 
     # Check if representation can be called
     def test_repr(self, pipe):
@@ -393,26 +393,26 @@ class Test_Pipeline_Init_Exceptions(object):
     def test_invalid_Emulator(self, root_working_dir, model_link):
         with pytest.raises(InputError):
             Pipeline(model_link, **root_working_dir,
-                     prism_file=prism_file_default, emul_type=InvalidEmulator)
+                     prism_par=prism_file_default, emul_type=InvalidEmulator)
 
     # Create a Pipeline object using not an Emulator class
     def test_no_Emulator(self, root_working_dir, model_link):
         with pytest.raises(InputError):
             Pipeline(model_link, **root_working_dir,
-                     prism_file=prism_file_default, emul_type=Pipeline)
+                     prism_par=prism_file_default, emul_type=Pipeline)
 
     # Create a Pipeline object using an improper ModelLink object
     def test_improper_ModelLink(self, root_working_dir):
         with pytest.raises(InputError):
             model_link = ImproperModelLink()
             Pipeline(model_link, **root_working_dir,
-                     prism_file=prism_file_default)
+                     prism_par=prism_file_default)
 
     # Create a Pipeline object using not a ModelLink object
     def test_no_ModelLink(self, root_working_dir):
         with pytest.raises(TypeError):
             Pipeline(np.array([1]), **root_working_dir,
-                     prism_file=prism_file_default)
+                     prism_par=prism_file_default)
 
     # Create a Pipeline object using invalid number of md_var
     def test_invalid_N_md_var(self, root_working_dir):
@@ -420,7 +420,7 @@ class Test_Pipeline_Init_Exceptions(object):
             InvalidNMdVarModelLink(model_parameters=model_parameters_3D,
                                    model_data=model_data_single)
         pipe = Pipeline(model_link, **root_working_dir,
-                        prism_file=prism_file_default)
+                        prism_par=prism_file_default)
         with pytest.raises(ShapeError):
             pipe._emulator._create_new_emulator()
 
@@ -430,7 +430,7 @@ class Test_Pipeline_Init_Exceptions(object):
             InvalidShapeMdVarModelLink(model_parameters=model_parameters_3D,
                                        model_data=model_data_single)
         pipe = Pipeline(model_link, **root_working_dir,
-                        prism_file=prism_file_default)
[…]
