+++ prism/_pipeline.py
@@ -68,7 +68,8 @@ class Pipeline(Projection, object):
 
     @docstring_substitute(paths=paths_doc_d)
     def __init__(self, modellink_obj, *, root_dir=None, working_dir=None,
-                 prefix=None, prism_file=None, emul_type=None, comm=None):
+                 prefix=None, prism_par=None, emul_type=None, comm=None,
+                 **kwargs):
         """
         Initialize an instance of the :class:`~Pipeline` class.
 
@@ -81,12 +82,16 @@ class Pipeline(Projection, object):
         Optional
         --------
         %(paths)s
-        prism_file : str or None. Default: None
-            String containing the absolute or relative path to the TXT-file
-            containing the *PRISM* parameters that need to be changed from
-            their default values. If a relative path is given, its path must be
-            relative to `root_dir` or the current directory. If *None*, no
-            changes will be made to the default parameters.
+        prism_par : array_like, dict, str or None. Default: None
+            A dict containing the values for the *PRISM* parameters that need
+            to be changed from their default values.
+            If array_like, dict(`prism_par`) must generate a dict with the
+            correct lay-out.
+            If str, the string is the absolute or relative path to the file
+            that contains the keys in the first column and the dict values in
+            the second column. If a relative path is given, its path must be
+            relative to `root_dir` or the current directory.
+            If *None*, no changes will be made to the default parameters.
         emul_type : :class:`~prism.emulator.Emulator` subclass or None.\
             Default: None
             The type of :class:`~prism.emulator.Emulator` to use in this
@@ -141,8 +146,19 @@ class Pipeline(Projection, object):
         if self._is_worker:
             set_base_logger(path.join(self._working_dir, 'prism_log.log'))
 
-        # Read in the provided parameter file
-        self._read_parameters(prism_file)
+        # Check if deprecated prism_file was provided
+        if 'prism_file' in kwargs.keys():
+            # Assign prism_file to prism_par
+            prism_par = kwargs['prism_file']
+
+            # Raise a FutureWarning
+            warn_msg = ("Input argument 'prism_file' is deprecated since "
+                        "v1.1.2 in favor of 'prism_par'. It will be removed "
+                        "entirely in v1.2.0.")
+            raise_warning(warn_msg, FutureWarning, logger, stacklevel=2)
+
+        # Read in the provided parameter dict/file
+        self._read_parameters(prism_par)
 
         # Initialize Emulator class
         # If emul_type is None, use default emulator
@@ -231,14 +247,34 @@ class Pipeline(Projection, object):
         str_repr.append("working_dir=%r" % (path.relpath(self._working_dir,
                                                          self._root_dir)))
 
-        # Add the prism_file representation if it is not default
-        if self._prism_file is not None:
-            if(path.splitdrive(self._prism_file)[0].lower() !=
-               path.splitdrive(pwd)[0].lower()):
-                str_repr.append("prism_file=%r" % (self._prism_file))
-            else:
-                str_repr.append("prism_file=%r"
-                                % (path.relpath(self._prism_file, pwd)))
+        # Add the prism_par representation if it is not default
+        if self._prism_dict:
+            # Make empty default dict
+            default_dict = sdict()
+
+            # Make empty list of prism_par representations
+            par_repr = []
+
+            # Add the default parameter dicts to it
+            default_dict.update(self._get_default_parameters())
+            default_dict.update(self._emulator._get_default_parameters())
+            default_dict.update(self._Projection__get_default_parameters())
+
+            # Loop over all items in prism_dict and check if it is default
+            for key, val in self._prism_dict.items():
[…]
