+++ docs/source/api/prism.modellink.rst
@@ -1,7 +1,5 @@
 ModelLink
 =========
-.. automodule:: prism.modellink.utils
-
 Classes
 -------
 .. toctree::
@@ -9,3 +7,8 @@ Classes
     prism.modellink.ModelLink
     prism.modellink.SineWaveLink
 
+Utilities
+---------
+.. autofunction:: prism.modellink.convert_data
+.. autofunction:: prism.modellink.convert_parameters
+.. autofunction:: prism.modellink.test_subclass
