### 2019-03-18 dbc44d7
Changed a forgotten 'prism_file' into a 'prism_par' in the tests.
