### 2019-03-18 6216379
Removed the use of 'emul_s_seq' in the _get_uni_impl_val() and _evaluate()-methods (as they were never used anyways).
