### 2019-03-19 e95137d
Fixed a bug where the custom gather()-method of the PRISM_Comm class did not create a buffer array with the same dtype as the object that was going to be gathered.
