### 2019-03-19 79f323e
Now that MPL v2.2.4 has been released, set the minimum requirement to that version.
