+++ requirements.txt
@@ -1,10 +1,9 @@
 e13tools>=0.4.7a1
 h5py>=2.8.0
 future>=0.17.0
-matplotlib>=2.2.0
+matplotlib>=2.2.4
 mlxtend>=0.9.1
-numpy>=1.12.0;python_version>="3"
-numpy>=1.12.0,<1.16.0;python_version=="2.7.*"
+numpy>=1.12.0
 scikit-learn>=0.19.1
 scipy>=1.0.0
 six>=1.11.0
