### 2019-03-19 7d22186
Now that MPL v2.2.4 has been released, set the minimum requirement to that version.
