+++ requirements.txt
@@ -1,7 +1,7 @@
 e13tools>=0.5.4
 h5py>=2.8.0
 hickle>=3.4.0
-matplotlib>=2.2.0
+matplotlib>=2.2.4
 mlxtend>=0.9.1
 numpy>=1.12.0
 scikit-learn>=0.19.1
