### 2019-03-20 81d7a72
Providing a Pipeline object to the get_lnpost_fn() and get_walkers() functions will raise an error if it does not use a default emulator.
