+++ setup.cfg
@@ -1,5 +1,5 @@
 [bdist_wheel]
-universal=0
+universal = 0
 
 [tool:pytest]
 addopts = --mpl --pep8 -v
