### 2019-03-27 544c5d8
Now that e13Tools automatically imports the proper MPI-module at root-level (in v0.5.6), changed several lines in the code to simply import MPI from e13Tools instead. Updated the requirement for e13Tools to v0.5.6. Removed the use of '*[list]' in the _evaluate_sam_set()-method, as it is completely pointless to do.
