+++ prism/__init__.py
@@ -29,8 +29,8 @@ import platform
 import warnings
 
 # Package imports
-from e13tools import (aux_char_list, compare_versions as _compare_versions,
-                      import_cmaps)
+from e13tools import (MPI as _MPI, aux_char_list,
+                      compare_versions as _compare_versions, import_cmaps)
 
 # PRISM imports
 from .__version__ import prism_version as __version__
@@ -52,36 +52,31 @@ __author__ = "Ellert van der Velden (@1313e)"
 # Import PRISM's custom cmaps
 import_cmaps(os.path.join(os.path.dirname(__file__), 'data'))
 
-# Check if MPI is being used
-try:
-    from mpi4py import MPI as _MPI
-except ImportError:
-    pass
-else:
-    # If so, perform some checks on controller if MPI_size > 1
-    if(_MPI.COMM_WORLD.Get_size() > 1 and _MPI.COMM_WORLD.Get_rank() == 0):
-        # Check if imported mpi4py package is at least 3.0.0
-        from mpi4py import __version__ as _mpi4py_version
-        if not _compare_versions(_mpi4py_version, '3.0.0'):
-            raise ImportError("mpi4py v%s detected. PRISM requires mpi4py "
-                              "v3.0.0 or later to work in MPI!"
-                              % (_mpi4py_version))
+# Check if MPI is being used and perform some checks on controller if size > 1
+if(_MPI.__package__ == 'mpi4py' and _MPI.COMM_WORLD.Get_size() > 1 and
+   _MPI.COMM_WORLD.Get_rank() == 0):
+    # Check if imported mpi4py package is at least 3.0.0
+    from mpi4py import __version__ as _mpi4py_version
+    if not _compare_versions(_mpi4py_version, '3.0.0'):
+        raise ImportError("mpi4py v%s detected. PRISM requires mpi4py "
+                          "v3.0.0 or later to work in MPI!"
+                          % (_mpi4py_version))
 
-        # Raise warning if OMP_NUM_THREADS is not set to 1
-        if(os.environ.get('OMP_NUM_THREADS') != '1'):
-            # Get platform-dependent string on how to set environment variable
-            set_str = ""
+    # Raise warning if OMP_NUM_THREADS is not set to 1
+    if(os.environ.get('OMP_NUM_THREADS') != '1'):
+        # Get platform-dependent string on how to set environment variable
+        set_str = ""
 
-            # Windows
-            if (platform.system().lower() == 'windows'):
-                set_str = " (\">set OMP_NUM_THREADS=1\")"
-            # Linux/MacOS-X
-            elif (platform.system().lower() in ('linux', 'darwin')):
-                set_str = " (\"$ export OMP_NUM_THREADS=1\")"
+        # Windows
+        if (platform.system().lower() == 'windows'):
+            set_str = " (\">set OMP_NUM_THREADS=1\")"
+        # Linux/MacOS-X
+        elif (platform.system().lower() in ('linux', 'darwin')):
+            set_str = " (\"$ export OMP_NUM_THREADS=1\")"
 
-            # Print warning message
-            warn_msg = ("Environment variable 'OMP_NUM_THREADS' is currently "
-                        "not set to 1 (%s), with MPI enabled. Unless this was "
-                        "intentional, it is advised to set it to 1%s."
-                        % (os.environ.get('OMP_NUM_THREADS'), set_str))
-            warnings.warn(warn_msg, RuntimeWarning, stacklevel=2)
+        # Print warning message
+        warn_msg = ("Environment variable 'OMP_NUM_THREADS' is currently "
+                    "not set to 1 (%s), with MPI enabled. Unless this was "
[…]
