### 2019-03-27 3a8e53f
Improved the calculation of the covariance to be slightly more efficient. The correlation length 'l_corr' is now also checked to be equal to or less than 1.
