### 2019-03-28 9931154
Added the mpi4pyd requirement to the package. Changed several instances to use the new mpi4pyd package instead of the e13Tools package.
