+++ requirements.txt
@@ -1,8 +1,9 @@
-e13tools>=0.5.6
+e13tools>=0.6.0
 h5py>=2.8.0
 hickle>=3.4.0
 matplotlib>=2.2.4
 mlxtend>=0.9.1
+mpi4pyd>=0.1.0
 numpy>=1.12.0
 scikit-learn>=0.19.1
 scipy>=1.0.0
