### 2019-04-03 917e501
Changed the reading of the PRISM version to no longer require the 'exec' function in the setup.py file (thanks to @abatten for the idea).
