+++ setup.py
@@ -30,21 +30,24 @@ with open('README.rst', 'r') as f:
 with open('requirements.txt', 'r') as f:
     requirements = f.read().splitlines()
 
-# Get the version
-prism_version = None
+# Read the __version__.py file
 with open('prism/__version__.py', 'r') as f:
-    exec(f.read())
+    vfile = f.read()
+
+# Obtain version from read-in __version__.py file
+vstr = "__version__ = "
+version = vfile.partition(vstr)[2].partition('\n')[0].replace("'", '').strip()
 
 # Setup function declaration
 setup(name='prism',
-      version=prism_version,
+      version=version,
       author="Ellert van der Velden",
       author_email="evandervelden@swin.edu.au",
       maintainer="1313e",
       description="PRISM: An alternative to MCMC for rapid analysis of models",
       long_description=long_description,
       download_url=("https://github.com/1313e/PRISM/archive/v%s.zip"
-                    % (prism_version)),
+                    % (version)),
       url="https://prism-tool.readthedocs.io",
       project_urls={
           'Documentation': "https://prism-tool.readthedocs.io/en/latest/",
