### 2019-04-03 06daf12
Also made the version change to the docs.
