### 2019-04-03 7f61569
Added missing __version__ to test_modellink.py
