### 2019-04-03 fc88aa0
Replaced a weird bug where characters were replaced (that have not been replaced at all).
