+++ prism/_internal.py
@@ -773,7 +773,7 @@ def convert_str_seq(seq):
 
 # List of auxiliary characters to be used in convert_str_seq()
 aux_char_list = ['(', ')', '[', ']', ',', "'", '"', '|', '/', '{', '}', '<',
-                 '>', '´', '¨', '`', '\\', '?', '!', '%', ';', '=', '$', '~',
+                 '>', 'Â´', 'Â¨', '`', '\\', '?', '!', '%', ';', '=', '$', '~',
                  '#', '@', '^', '&', '*']
 
 
