### 2019-04-03 dfdab4a
Since e13Tools v0.6.1 uses aux_char_set now instead of aux_char_list, changed all instances of using those to the correct one.
