### 2019-04-03 965708b
Removed unused function in the test_modellink.py file.
