### 2019-04-04 3315f8c
Started using regex for single-sourcing the version string.
