### 2019-04-04 13a4290
Started using regex for single-sourcing the version string.
