+++ setup.py
@@ -19,6 +19,7 @@ from __future__ import absolute_import, with_statement
 
 # Built-in imports
 from codecs import open
+import re
 
 # Package imports
 from setuptools import find_packages, setup
@@ -38,12 +39,11 @@ with open('prism/__version__.py', 'r') as f:
     vfile = f.read()
 
 # Obtain version from read-in __version__.py file
-vstr = "__version__ = "
-version = vfile.partition(vstr)[2].partition('\n')[0].replace("'", '').strip()
+version = re.search(r"^_*version_* = ['\"]([^'\"]*)['\"]", vfile, re.M)
 
 # Setup function declaration
 setup(name='prism',
-      version=version,
+      version=version.group(1),
       author="Ellert van der Velden",
       author_email="evandervelden@swin.edu.au",
       maintainer="1313e",
