### 2019-04-04 420a79c
Make sure that use of regular expression is a one-liner.
