+++ setup.py
@@ -36,14 +36,14 @@ with open('requirements.txt', 'r') as f:
 
 # Read the __version__.py file
 with open('prism/__version__.py', 'r') as f:
-    vfile = f.read()
+    vf = f.read()
 
 # Obtain version from read-in __version__.py file
-version = re.search(r"^_*version_* = ['\"]([^'\"]*)['\"]", vfile, re.M)
+version = re.search(r"^_*version_* = ['\"]([^'\"]*)['\"]", vf, re.M).group(1)
 
 # Setup function declaration
 setup(name='prism',
-      version=version.group(1),
+      version=version,
       author="Ellert van der Velden",
       author_email="evandervelden@swin.edu.au",
       maintainer="1313e",
