### 2019-04-04 92a3d72
Make sure that use of regular expression is a one-liner.
