### 2019-04-04 d247590
Add project URLs to v1.0.x branch.
