### 2019-04-04 f216bb5
The analyze()-method will now raise a warning if the next iteration cannot be constructed. It is no longer possible to construct an iteration if less than n_cross_val plausible samples will be used for it.
