+++ prism/tests/test_pipeline_emulator_projection.py
@@ -25,7 +25,7 @@ from sortedcontainers import SortedDict as sdict
 
 # PRISM imports
 from prism._emulator import Emulator
-from prism._internal import RequestError, check_instance
+from prism._internal import RequestError, RequestWarning, check_instance
 from prism._pipeline import Pipeline
 from prism._projection import Projection
 from prism.modellink import ModelLink
@@ -49,6 +49,7 @@ pytestmark = pytest.mark.filterwarnings(
 model_data_single = path.join(dirpath, 'data/data_gaussian_single.txt')
 prism_file_default = path.join(dirpath, 'data/prism_default.txt')
 prism_file_impl = path.join(dirpath, 'data/prism_impl.txt')
+prism_file_n_cross_val = path.join(dirpath, 'data/prism_n_cross_val.txt')
 model_parameters_2D = path.join(dirpath, 'data/parameters_gaussian_2D.txt')
 model_parameters_3D = path.join(dirpath, 'data/parameters_gaussian_3D.txt')
 
@@ -763,6 +764,16 @@ class Test_Pipeline_Request_Exceptions(object):
 
     # Create a universal Pipeline object for testing request exceptions
     @pytest.fixture(scope='class')
+    def pipe_n_cross_val(self, tmpdir_factory):
+        tmpdir = tmpdir_factory.mktemp('test_request_exceptions_n_cross_val')
+        root_dir = path.dirname(tmpdir.strpath)
+        working_dir = path.basename(tmpdir.strpath)
+        model_link = GaussianLink2D()
+        return(Pipeline(model_link, root_dir=root_dir, working_dir=working_dir,
+                        prism_par=prism_file_n_cross_val))
+
+    # Create a universal Pipeline object for testing request exceptions
+    @pytest.fixture(scope='class')
     def pipe_default(self, tmpdir_factory):
         tmpdir = tmpdir_factory.mktemp('test_request_exceptions_default')
         root_dir = path.dirname(tmpdir.strpath)
@@ -821,7 +832,7 @@ class Test_Pipeline_Request_Exceptions(object):
 
     # Try to call an iteration that cannot be finished
     def test_break_call(self, pipe_impl):
-        with pytest.raises(RequestError):
+        with pytest.raises(RequestError), pytest.warns(RequestWarning):
             pipe_impl(1)
 
     # Try to construct an iteration with no plausible regions
@@ -845,6 +856,13 @@ class Test_Pipeline_Request_Exceptions(object):
         with pytest.raises(RequestError):
             pipe_impl._emulator._retrieve_parameters()
         pipe_impl._emulator._emul_type = 'default'
+
+    # Try to construct an iteration that has less than n_cross_val samples
+    def test_n_cross_val_construction(self, pipe_n_cross_val):
+        with pytest.warns(RequestWarning):
+            pipe_n_cross_val.construct(1)
+        with pytest.raises(RequestError):
+            pipe_n_cross_val.construct(2)
 
 
 # Pytest for Pipeline class internal exception handling
