### 2019-04-04 12987a1
Fixed a bug in the tests caused by previous cherry-picking.
