### 2019-04-24 #1229 @1313e
@crawfordsm @arfon @arokem
Alright, so I am currently in the process of revising the AAS Journals submission.
One of the things that still needs to be done, is linking the JOSS submission and the AAS Journals submission together.
However, as neither submission will get accepted before the other, no DOI exists for either submission.
This makes it hard for me to cite one submission in the other, which is necessary for both.

So, therefore, I wanted to ask: Any ideas on how I should cite either submission in the other?

PS: I coincidentally went on a three week holiday the day that I received the referee report from AAS Journals, which is why it took until now to get to it.
