### 2019-04-24 #1229 @arfon
@1313e - we actually know ahead of time what the JOSS DOI will be - in this case it's https://doi.org/10.21105/joss.01229 (this won't resolve yet though).

I also believe that AAS _should_ know their DOI towards the end of their editorial process. @crawfordsm - is that the case?
