### 2019-04-24 #1229 @1313e
@arfon Right, I forgot that the DOI for JOSS submissions is actually really easy.
However, I still wanted to raise this, as I think it is a good idea to discuss how to proceed.
