### 2019-04-24 #1229 @arfon
> However, I still wanted to raise this, as I think it is a good idea to discuss how to proceed.

Yes, definitely worth discussing! @crawfordsm and I are overdue producing some documentation for authors such as yourself describing how all of this works 😄
