### 2019-04-26 f9e0894
Added statement that PRISM is based on the work done by Vernon et al. (2010).
