### 2019-04-26 #1229 @1313e
Also, if I am looking at the volume and issue number logic of previously accepted papers, I am assuming that this one is going to be `4(37)` (as it is probably not going to be accepted this month).
I am asking as I want to include the full citation to the JOSS paper in the AAS Journals submission, and I can pretty much guess what the entire bibtex entry is going to look like.
