### 2019-04-29 55652c6
Removed unnecessary parentheses around first reference.
