### 2019-04-30 5fd55fc
Made sure that the FeatureWarning for 'prism_file' is now solely raised on the controller. Calling 'details' when the emulator has not been built yet, no longer logs an error that was caught.
