### 2019-05-02 0a2b618
Added the 'show_cuts' parameter to the project()-method, which determines whether to show all implausibility cut-off lines in 2D projections or only the first one. Added the 'line_kwargs_cut' parameter as well, which gives the keywords arguments for drawing the cut-off lines. Renamed 'line_kwargs' to 'line_kwargs_est' to accommodate for the change above. Added some more comments to the draw-methods for projections.
