+++ prism/_projection.py
@@ -98,6 +98,10 @@ class Projection(object):
             If 'row'/'horizontal', the subplots are positioned on a single row.
             If 'col'/'column'/'vertical', the subplots are positioned on a
             single column.
+        show_cuts : bool. Default: False
+            If `figure` is *True* and `proj_type` is not '3D', whether to show
+            all implausibility cut-offs in the 2D projections (*True*) or only
+            the first cut-off (*False*).
         smooth : bool. Default: False
             Controls what to do if a grid point contains no plausible samples,
             but does contain a minimum implausibility value below the first
@@ -140,10 +144,15 @@ class Projection(object):
             (bottom/right) plot in the 3D projection figures. It takes all
             arguments that can be provided to the
             :func:`~matplotlib.pyplot.hexbin` function.
-        line_kwargs : dict. Default: {'linestyle': '--', 'color': 'grey'}
+        line_kwargs_est : dict. Default: {'linestyle': '--', 'color': 'grey'}
             Dict of keyword arguments to be used for drawing the parameter
             estimate lines in both plots. It takes all arguments that can be
             provided to the :func:`~matplotlib.pyplot.draw` function.
+        line_kwargs_cut : dict. Default: {'color': 'r'}
+            Dict of keyword arguments to be used for drawing the implausibility
+            cut-off line(s) in the top/left plot in the 2D projection figures.
+            It takes all arguments that can be provided to the
+            :func:`~matplotlib.pyplot.draw` function.
 
         Returns (if `figure` is *False*)
         --------------------------------
@@ -311,8 +320,9 @@ class Projection(object):
         # Obtain name of this projection hypercube
         hcube_name = self.__get_hcube_name(hcube)
 
-        # Make abbreviation for first implausibility cut-off value
+        # Make abbreviation for implausibility cut-off values
         impl_cut = self._impl_cut[self.__emul_i][0]
+        impl_cuts = self._impl_cut[self.__emul_i]
 
         # Start logger
         logger = getCLogger('PROJECTION')
@@ -392,25 +402,45 @@ class Projection(object):
             f.suptitle(r"%s. Projection (%s)" % (self.__emul_i, hcube_name),
                        fontsize='xx-large')
 
+            # MINIMUM IMPLAUSIBILITY PLOT
             # Plot minimum implausibility
             ax0.plot(x, y_min, **self.__impl_kwargs_2D)
             ax0_rng = [*self._modellink._par_rng[par], 0, 1.5*impl_cut]
             ax0.axis(ax0_rng)
+
+            # Draw parameter estimate line
             if self._modellink._par_est[par] is not None:
                 draw_textline(r"", x=self._modellink._par_est[par], ax=ax0,
-                              line_kwargs=self.__line_kwargs)
-            draw_textline(r"", y=impl_cut, ax=ax0, line_kwargs={'color': 'r'})
+                              line_kwargs=self.__line_kwargs_est)
+
+            # Draw implausibility cut-off line(s)
+            if self.__show_cuts:
+                # If all lines are requested, draw them
+                for cut in impl_cuts:
+                    draw_textline(r"", y=cut, ax=ax0,
+                                  line_kwargs=self.__line_kwargs_cut)
+            else:
+                # Else, draw the first cut-off line
+                draw_textline(r"", y=impl_cut, ax=ax0,
+                              line_kwargs=self.__line_kwargs_cut)
+
+            # Set axes and label
             ax0.axis(ax0_rng)
             ax0.set_ylabel("Min. Implausibility", fontsize='large')
 
[…]
