### 2019-05-04 #1229 @1313e
@crawfordsm Now that the AAS Journals submission has been accepted, how do we proceed here?
