### 2019-05-06 #1229 @arfon
> @arfon I heard there are some problems with cross-refs at the moment?

Yeah, I suspect this will get fixed tomorrow. For now, we're not able to deposit with Crossref.
