### 2019-05-06 #1229 @1313e
@arfon @crawfordsm Has it already been decided how the JOSS and AAS Journals submissions are going to be linked? Now that the AAS Journals submission has been accepted as well, I guess it is time to link them together (which can be done while cross-refs are not working anyways)?

PS: I remember a friend of mine showing me something from JOSS last Friday, that looked like some special submission page had been set up for AAS Journals and JOSS submissions. I however am unable to find it again, so I am not sure what it was.
