### 2019-05-06 e756b20
Set self._worker_mode to 1 in the _listen_for_calls()-method instead of while entering the worker_mode context manager.
