### 2019-05-06 #1229 @arfon
> @arfon @crawfordsm Has it already been decided how the JOSS and AAS Journals submissions are going to be linked? Now that the AAS Journals submission has been accepted as well, I guess it is time to link them together (which can be done while cross-refs are not working anyways)?

We're going to as you to make sure that both papers cite each other. Right now, we don't know what the AAS DOI will be so having the JOSS paper cite the AAS paper isn't likely possible for another few days.

> PS: I remember a friend of mine showing me something from JOSS last Friday, that looked like some special submission page had been set up for AAS Journals and JOSS submissions. I however am unable to find it again, so I am not sure what it was.

We don't have anything like this sorry. Perhaps you're thinking of this http://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing ?
