### 2019-05-07 #1229 @1313e
> We don't have anything like this sorry. Perhaps you're thinking of this http://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing ?

No, I remember there being some kind of form with the logo of JOSS covering the entire background of it.
Problem is that I cannot ask the person who showed it to me, as he is on a holiday.
I might be wrong though, perfectly possible.

However, I am assuming that the DOI for AAS Journals will be known when the paper is actually published.
So, that would mean waiting for at least another month.
