### 2019-05-08 780215d
Made some small changes to docstrings concerning MPI intra-communicators.
