### 2019-05-08 b42aa72
Added AAS Journals submission citation with DOI to JOSS submission.
