+++ joss_paper/paper.bib
@@ -12,7 +12,7 @@
    author = {{Bower}, R.~G. and {Vernon}, I. and {Goldstein}, M. and {Benson}, A.~J. and 
 	{Lacey}, C.~G. and {Baugh}, C.~M. and {Cole}, S. and {Frenk}, C.~S.},
     title = "{The parameter space of galaxy formation}",
-  journal = {\mnras},
+  journal = {Monthly Notices of the Royal Astronomical Society},
 archivePrefix = "arXiv",
    eprint = {1004.0711},
  keywords = {galaxies: formation, galaxies: luminosity function, mass function},
@@ -151,7 +151,7 @@ isbn="978-1-4612-2290-3"
    author = {{Mutch}, S.~J. and {Geil}, P.~M. and {Poole}, G.~B. and {Angel}, P.~W. and 
 	{Duffy}, A.~R. and {Mesinger}, A. and {Wyithe}, J.~S.~B.},
     title = "{Dark-ages reionization and galaxy formation simulation - III. Modelling galaxy formation and the epoch of reionization}",
-  journal = {\mnras},
+  journal = {Monthly Notices of the Royal Astronomical Society},
 archivePrefix = "arXiv",
    eprint = {1512.00562},
  keywords = {galaxies: formation, galaxies: high redshift, dark ages, reionization, first stars},
@@ -170,4 +170,19 @@ archivePrefix = "arXiv",
   year =      {2006},
   publisher = {USA: Trelgol Publishing},
   url = "http://www.numpy.org/"
+}
+
+@ARTICLE{PRISM,
+   author = {{van der Velden}, E. and {Duffy}, A.~R. and {Croton}, D. and 
+	{Mutch}, S.~J. and {Sinha}, M.},
+    title = "{Model dispersion with PRISM; an alternative to MCMC for rapid analysis of models}",
+  journal = {The Astrophysical Journal Supplement Series},
+archivePrefix = "arXiv",
+   eprint = {1901.08725},
+ primaryClass = "astro-ph.IM",
+ keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics},
+     year = 2019,
+      doi = {10.3847/1538-4365/ab1f7d},
+   adsurl = {http://adsabs.harvard.edu/abs/2019arXiv190108725V},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
