### 2019-05-08 #1229 @1313e
@arfon @crawfordsm Well, since I got the DOI for the ApJS manuscript this morning, I have added it to the JOSS manuscript.
I guess there are some other things that need to be changed or added, as currently it is not very clear that both manuscripts are somewhat linked.
