### 2019-05-08 #1229 @arfon
> @arfon @crawfordsm Well, since I got the DOI for the ApJS manuscript this morning, I have added it to the JOSS manuscript.

Excellent. You might want to add the issue/volume numbers when we have them too.

> I guess there are some other things that need to be changed or added, as currently it is not very clear that both manuscripts are somewhat linked.

Right. I'm talking with AAS publishing about finding a way to show these are linked in some way. Watch this space - I should have an update on this soon. Also, thanks for your patience (and being the first to do this) - as you can tell, we have some things to work through with AAS here 😸
