### 2019-05-10 75aa3da
Moved the GaussianLink class properties to match the lay-out used for other classes in PRISM.
