### 2019-05-15 2f083a9
The regression function now solely contains polynomial terms with coefficients that are significantly different from zero (unless none of them are). This removes having polynomial terms with 1e-16 coefficients, adding nothing to the accuracy but still slowing down the calculations.
