### 2019-05-20 #1229 @arfon
> @arfon Any news/updates?

@crawfordsm - any news from AAS on this paper?

> Right. I'm talking with AAS publishing about finding a way to show these are linked in some way. Watch this space - I should have an update on this soon. Also, thanks for your patience (and being the first to do this) - as you can tell, we have some things to work through with AAS here 😸

BTW, we've decided that the JOSS papers won't need page/issue/volume numbers for the AAS papers as long as they have the correct DOI for the AAS paper in the citation.
