### 2019-05-21 #1229 @1313e
@arfon So, is there going to be something like a unique/specific identifier going to be added to both papers (or something like a note) that states the papers are linked?
