### 2019-05-22 180a870
The minimum implausibility colorbar now shows that it extends beyond the first cut-off value.
