### 2019-05-29 6c0b73e
Added new PRISM logo and acknowledgements to the docs.
