### 2019-05-29 db1d80c
Added new PRISM logo and acknowledgements to the docs.
