+++ docs/source/conf.py
@@ -145,7 +145,9 @@ html_theme_options = {
     'collapse_navigation': False,
     'sticky_navigation': False,
     'includehidden': False,
-    'titles_only': False
+    'titles_only': False,
+    'logo_only': True,
+    'display_version': False
 }
 
 # Title formatting
@@ -174,7 +176,8 @@ html_use_opensearch = 'https://prism-tool.readthedocs.io/en/v1.0.x'
 
 html_split_index = True
 
-html_favicon = 'py.png'
+html_logo = 'PRISM_Logo.png'
+html_favicon = 'PRISM_Icon.ico'
 
 html_baseurl = 'https://prism-tool.readthedocs.io/en/v1.0.x'
 
