### 2019-05-29 0528e11
Added PRISM logo to the README.rst file.
