### 2019-05-29 3c82ae3
Added PRISM logo to the README.rst file.
