+++ README.rst
@@ -5,6 +5,10 @@ Starting with v1.1.0, *PRISM* no longer supports Python 2.7, and this branch wil
 
 ----
 
+.. image:: ./logo/PRISM_transparent_Logo1.svg
+    :width: 400 px
+    :align: center
+
 |PyPI| |Python| |License| |Travis| |AppVeyor| |ReadTheDocs| |Coverage|
 
 Model dispersion with *PRISM*; an alternative to MCMC for rapid analysis of models
