### 2019-05-29 07c9ea0
Removed the PRISM_Comm class definition (as it is now fully absorbed in the mpi4pyd package).
