+++ prism/_internal.py
@@ -32,8 +32,8 @@ from pkg_resources import get_distribution
 from prism.__version__ import __version__, compat_version
 
 # All declaration
-__all__ = ['CFilter', 'FeatureWarning', 'PRISM_Comm', 'PRISM_Logger',
-           'RFilter', 'RequestError', 'RequestWarning', 'check_compatibility',
+__all__ = ['CFilter', 'FeatureWarning', 'PRISM_Logger', 'RFilter',
+           'RequestError', 'RequestWarning', 'check_compatibility',
            'check_vals', 'get_PRISM_File', 'get_formatter', 'get_handler',
            'get_info', 'getCLogger', 'getLogger', 'getRLogger', 'move_logger',
            'np_array', 'set_base_logger']
@@ -72,193 +72,6 @@ class FeatureWarning(FutureWarning):
     """
 
     pass
-
-
-# Make a custom MPI.Comm class that uses a special broadcast method
-# TODO: Combine this class with dummyMPI into a separate dedicated package?
-class PRISM_Comm(object):
-    """
-    Custom :class:`~mpi4pyd.MPI.Intracomm` class that automatically makes use
-    of the :class:`~numpy.ndarray` buffers when using communications. Is
-    functionally the same as the provided `comm` for everything else.
-
-    Optional
-    --------
-    comm : :obj:`~mpi4pyd.MPI.Intracomm` object or None. Default: None
-        The MPI intra-communicator to use in this :class:`~PRISM_Comm`
-        instance.
-        If *None*, use :obj:`MPI.COMM_WORLD` instead.
-
-    """
-
-    def __init__(self, comm=None):
-        # If comm is None, use MPI.COMM_WORLD
-        if comm is None:
-            comm = MPI.COMM_WORLD
-        # Else, raise error if provided comm is not an MPI intra-communicator
-        elif not isinstance(comm, MPI.Intracomm):
-            raise TypeError("Input argument 'comm' must be an instance of the "
-                            "MPI.Intracomm class!")
-
-        # Bind provided communicator
-        self._comm = comm
-        self._rank = self._comm.Get_rank()
-        self._size = self._comm.Get_size()
-
-    # Override getattr property to use self._comm attributes if necessary
-    def __getattribute__(self, name):
-        try:
-            return(super().__getattribute__(name))
-        except AttributeError:
-            return(getattr(self._comm, name))
-
-    # Override __dir__ attribute to use the one from self._comm
-    def __dir__(self):
-        return(dir(self._comm))
-
-    # Specialized bcast function that automatically makes use of NumPy buffers
-    def bcast(self, obj, root):
-        """
-        Special broadcast method that automatically uses the appropriate method
-        (:meth:`~MPI.Intracomm.bcast` or :meth:`~MPI.Intracomm.Bcast`)
-        depending on the type of the provided `obj`.
-
-        Parameters
-        ----------
-        obj : :obj:`~numpy.ndarray` or object
-            The object to broadcast to all MPI ranks.
-            If :obj:`~numpy.ndarray`, use :meth:`~MPI.Intracomm.Bcast`.
-            If not, use :meth:`~MPI.Intracomm.bcast` instead.
-        root : int
-            The MPI rank that broadcasts `obj`.
-
-        Returns
-        -------
-        obj : object
-            The broadcasted `obj`.
-
-        """
-
-        # Sender
-        if(self._rank == root):
-            # Check if provided object is a NumPy array
-            if isinstance(obj, np.ndarray):
-                # If so, send shape and dtype of the NumPy array
-                self._comm.bcast(['NumPy ndarray', [obj.shape, obj.dtype]],
-                                 root=root)
-
-                # Then send the NumPy array as a buffer object
[…]
