### 2019-05-29 074d2b4
Added forgotten change to using new get_BufferComm_obj()-function from mpi4pyd.
