### 2019-05-31 d8ba2bd
As mpi4pyd changed the name to HybridComm, updated the code and requirements.
