+++ prism/modellink/utils.py
@@ -18,7 +18,7 @@ import warnings
 # Package imports
 from e13tools import InputError
 from e13tools.utils import check_instance, convert_str_seq
-from mpi4pyd.MPI import get_BufferComm_obj
+from mpi4pyd.MPI import get_HybridComm_obj
 import numpy as np
 from numpy.random import rand
 from sortedcontainers import SortedDict as sdict
@@ -361,7 +361,7 @@ def test_subclass(subclass, *args, **kwargs):
                              "argument %r!" % (modellink_obj._name, arg))
 
     # Set MPI intra-communicator
-    comm = get_BufferComm_obj()
+    comm = get_HybridComm_obj()
 
     # Obtain random sam_set on controller
     if not comm._rank:
