### 2019-06-06 9b1b1ff
Mock data can now also be requested for a specified parameter set. The _check_sam_set()-method of the ModelLink class now makes sure that 1D sample sets are checked properly. Fixed up some stuff in the docstrings.
