+++ prism/utils/tests/test_mcmc.py
@@ -38,6 +38,7 @@ def pipe(tmpdir_factory):
     working_dir = path.basename(tmpdir.strpath)
     prism_file = path.join(dirpath, 'data/prism_default.txt')
     modellink_obj = GaussianLink2D()
+    np.random.seed(0)
     pipeline_obj = Pipeline(modellink_obj, root_dir=root_dir,
                             working_dir=working_dir, prism_par=prism_file)
     pipeline_obj.construct(analyze=False)
