### 2019-06-07 9ecb781
Some small updates to various docstrings.
