### 2019-06-07 fd6ca34
Now that e13Tools has online documentation, added it to the intersphinx mapping list.
