+++ docs/source/conf.py
@@ -58,7 +58,8 @@ intersphinx_mapping = {'python': ('https://docs.python.org/3', None),
                        'numpy': ('https://docs.scipy.org/doc/numpy', None),
                        'mpi4py': ('https://mpi4py.readthedocs.io/en/stable', None),
                        'matplotlib': ('https://matplotlib.org', None),
-                       'h5py': ('https://h5py.readthedocs.io/en/stable', None)}
+                       'h5py': ('https://h5py.readthedocs.io/en/stable', None),
+                       'e13tools': ('https://e13tools.readthedocs.io/en/latest', None)}
 
 # Autodoc configuration
 autodoc_default_options = {'members': None,
@@ -236,7 +237,7 @@ man_pages = [
 #  dir menu entry, description, category)
 texinfo_documents = [
     (master_doc, project, html_title,
-     author, project, '"Probabilistic Regression Instrument for Simulating Models"',
+     author, project, "An alternative to MCMC for rapid analysis of models",
      'Miscellaneous'),
 ]
 
