### 2019-06-07 94dfa53
Now that e13Tools has online documentation, added it to the intersphinx mapping list.
