### 2019-06-08 #1229 @arfon
@1313e - am I correct that the AAS DOI for this paper will be `10.3847/1538-4365/ab1f7d`?

If so, please add the following fields to the top of your `paper.md` file:

```
aas-doi: 10.3847/1538-4365/ab1f7d
aas-journal: Astrophysical Journal
```
