### 2019-06-09 #1229 @1313e
@arfon Sure. However, shouldn't it be "Astrophysics Journal Supplement Series"?
