### 2019-06-09 23634b5
Made some adjustments to the JOSS paper.
