### 2019-06-09 #1229 @arfon
Yes, please change the line: `aas-journal: Astrophysical Journal` to `aas-journal: Astrophysics Journal Supplement Series`
