### 2019-06-10 ff8952e
JOSS paper now mentions ApJS instead of ApJ.
