### 2019-06-10 #1229 @arfon
@1313e - I'm going to go ahead and accept this paper now as I believe the AAS paper is very close to being published too.

Apologies for the back and forth on this submission - you're the first author to follow the dual-publishing workflow in AAS/JOSS and we've been finding (and fixing) a few challenges along the way. Your patience has been much appreciated!
