### 2019-06-10 #1229 @arfon
> @arfon Will you also make sure that my last name is properly written in the published version?

Certainly. Can you verify that it looks correct in the pdf here? http://joss.theoj.org/papers/f31acc7b1be30757526442034888aabf
