### 2019-06-16 #1229 @1313e
@arfon I was actually just looking at the paper, and noticed that the citations are not displayed properly. The first in every sequence is, but the other ones are not.
As far as I know, this was not the case with the automatically compiled versions.
Did I write the citations wrong?
