### 2019-06-16 #1229 @1313e
@arfon Well, the second sentence of the third paragraph shows up as "PRISM uses the Bayes linear approach (Goldstein & Wooff, 2007), emulation technique (Craig, Goldstein, Seheult, & Smith, 1996, p. @Craig97) and history matching (Raftery, Givens, & Zeh, 1995, pp. @Craig96, @Craig97)...".
I have the feeling that a few citations here were not actually formatted.
