### 2019-06-16 #1229 @arfon
Ah, right. Sorry about that. The correct syntax for multiple citations is to separate them with semi-colons (e.g. `[@Raftery95; @Craig96; @Craig97]`). I've fixed up the paper locally in https://github.com/openjournals/joss-papers/blob/a66fbee618680adc4825d723edcbb40165e1dda0/joss.01229/10.21105.joss.01229.pdf and it should be correct now on the live site (note it can take a few hours for the updated PDF to show properly because of caching on our website).
