### 2019-06-16 #1229 @1313e
@arfon Thanks for that, but I think you will have to manually set my last name in the footer again...
