### 2019-06-17 #1229 @1313e
> The paper is now fixed.
>
> Can you check that this looks OK? https://www.doi2bib.org/bib/10.21105/joss.01229

That link does not resolve.
