### 2019-06-17 #1229 @arfon
Weird works fine for me. Can you try going to https://www.doi2bib.org/ and putting in `10.21105/joss.01229`.

If that doesn't work, the output from the service (i.e. the bibtex entry from the Crossref metadata) is:

```
@article{vanderVelden2019,
  doi = {10.21105/joss.01229},
  url = {https://doi.org/10.21105/joss.01229},
  year = {2019},
  month = jun,
  publisher = {The Open Journal},
  volume = {4},
  number = {38},
  pages = {1229},
  author = {Ellert van der Velden},
  title = {Model dispersion with {PRISM}$\mathsemicolon$ an alternative to {MCMC} for rapid analysis of models},
  journal = {Journal of Open Source Software}
}
```

☝️does that look OK?
