### 2019-06-17 #1229 @1313e
Uhm, my last name should be in braces, as otherwise it will not be formatted correctly.
Besides that, it looks good.

It does seem though that for my paper, my name is at least written correctly.
The JOSS paper for the Hickle package (21105/joss.01115) does not render it properly on that website, giving it as "Ellert Van Velden" again.
