### 2019-06-17 #1229 @arfon
OK, thanks. I don't have any control over where https://www.doi2bib.org puts the braces but the [metadata looks correct](https://github.com/openjournals/joss-papers/blob/master/joss.01229/10.21105.joss.01229.crossref.xml#L38-L42).

> The JOSS paper for the Hickle package (21105/joss.01115) does not render it properly on that website, giving it as "Ellert Van Velden" again.

OK, that's fixed for `10.21105/joss.01115` in https://github.com/openjournals/joss-papers/blob/master/joss.01115/10.21105.joss.01115.crossref.xml#L43-L47 but it will take a few hours to update on https://www.doi2bib.org .
