### 2019-06-17 #1229 @1313e
Alright, thanks.
I know from experience that having no braces can sometimes mess it up when using `natbib` for example, but I guess that if the metadata is correct, such a thing would not be hard to notice.
