### 2019-07-05 #1229 @arfon
> @arfon This is not that big of a deal, but I just realized that the footer in the JOSS paper says 'van der Velden et al.', even though I am the sole author on the paper.

Sorry about that. This should be fixed now. Note, the new pdf can take a few hours to show up sometimes because of browser caching.
